### 2021-04-05 #3146 @whedon
[PRE REVIEW]: Sapsan: Framework for Supernovae Turbulence Modeling with Machine Learning
**Submitting author:** @pikarpov-lanl (<a href="http://orcid.org/0000-0003-4311-8490">Platon I. Karpov</a>)
**Repository:** <a href="https://github.com/pikarpov-LANL/Sapsan" target ="_blank">https://github.com/pikarpov-LANL/Sapsan</a>
**Version:** v0.2.5
