### 2021-04-05 #3146 @danielskatz
@pikarpov-LANL - Given the comment in your README: "currently Sapsan is in alpha, but we are actively working on it and introduce new feature on a daily basis" and given that JOSS is looking for papers that are complete, I suggest this submission be withdrawn until it is more stable and more mature.  Does this seem ok to you?  If so, I will mark it as withdrawn.
