### 2021-04-07 #3146 @pikarpov-LANL
Hi Daniel,

This is an outdated note, left behind before I had any physics-informed methods, analytical tools, and wiki. As it stands, Sapsan is full-featured and ready for general use as outlined in the paper. I have removed that note from README.

Thank you for bringing it to my attention!

Plato
