### 2021-04-08 #3146 @dfm
@pikarpov-LANL: Can you make these small changes:

1. Fix the ORCID in the manuscript
2. Add the [suggested DOI](https://github.com/openjournals/joss-reviews/issues/3146#issuecomment-815724155): `10.1109/tkde.2017.2720168`
3. Add an [explicit "Statement of need" section](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) to the paper
