+++ docs/joss/paper.bib
@@ -179,11 +179,12 @@ archivePrefix = {arXiv},
 archivePrefix = {arXiv},
        eprint = {1612.08544},
  primaryClass = {cs.LG},
+ 	  doi = {10.1109/tkde.2017.2720168},
        adsurl = {https://ui.adsabs.harvard.edu/abs/2016arXiv161208544K},
       adsnote = {Provided by the SAO/NASA Astrophysics Data System}
 }
 
-﻿@Article{Eyink2013,
+@Article{Eyink2013,
 author={Eyink, Gregory
 and Vishniac, Ethan
 and Lalescu, Cristian
