### 2021-04-21 #3199 @whedon
[REVIEW]: Sapsan: Framework for Supernovae Turbulence Modeling with Machine Learning
**Submitting author:** @pikarpov-lanl (<a href="http://orcid.org/0000-0003-4311-8490">Platon I. Karpov</a>)
**Repository:** <a href="https://github.com/pikarpov-LANL/Sapsan" target ="_blank">https://github.com/pikarpov-LANL/Sapsan</a>
**Version:** v0.4.6
