+++ sapsan/examples/GUI/st_config.txt
@@ -3,11 +3,10 @@ experiment name = CNN experiment
 mlflow_host = localhost
 mlflow_port = 9000
 edit_port = 8601
-axis = 3
-checkpoint_data_size = 32
-sample_to = 16
-batch_size = 8
-n_epochs = 1
+input_size = (32,32,32)
+sample_to = (16,16,16)
+batch_size = (8,8,8)
+n_epochs = 5
 patience = 10
 min_delta = 1e-5
 path = ../data/t{checkpoint:1.0f}/{feature}_dim32_fm15.h5
