+++ sapsan/utils/plot.py
@@ -205,8 +205,8 @@ def model_graph(model, shape: np.array, transforms = None):
         shape[1] = 1
         warnings.warn("shape was changed to %s to draw a model graph."%str(shape))
     
-    if len(shape) == 5: unit_input = torch.zeros([shape[0], 1, shape[2], shape[3], shape[4]])
-    elif len(shape) == 4: unit_input = torch.zeros([shape[0], 1, shape[2], shape[3]])
+    if len(shape) == 5: unit_input = torch.zeros(tuple(shape))
+    elif len(shape) == 4: unit_input = torch.zeros(tuple(shape))
     else: raise ValueError('Input shape can be either of 2D or 3D data')
         
     graph = hl.build_graph(model, unit_input, transforms = transforms)
