### 2021-04-30 03e285c
GUI fixes (#96)
* Updated the GUI
	- brought up to date with the latest backend
