+++ sapsan/examples/GUI/st_experiments.py
@@ -9,8 +9,7 @@ sys.path.append(str(Path.home())+"/Sapsan/")
 
 from sapsan.lib.backends.fake import FakeBackend
 from sapsan.lib.backends.mlflow import MLflowBackend
-from sapsan.lib.data.hdf5_dataset import HDF5Dataset
-from sapsan.lib.data import EquidistanceSampling
+from sapsan.lib.data import HDF5Dataset, EquidistantSampling, flatten
 from sapsan.lib.estimator import CNN3d, CNN3dConfig
 from sapsan.lib.estimator.cnn.spacial_3d_encoder import CNN3dModel
 from sapsan.lib.experiments.evaluate import Evaluate
@@ -59,11 +58,7 @@ def intro():
         > ## **Purpose**
         
         > Sapsan takes out all the hard work from data preparation and analysis in turbulence 
-        > and astrophysical applications, leaving you focused on ML model design, layer by layer.
-        
-
-        Note: currently Sapsan is in alpha, but we are actively working on it and introduce new 
-        feature on a daily basis.        
+        > and astrophysical applications, leaving you focused on ML model design, layer by layer.             
 
         **👈 Select an experiment from the dropdown on the left** to see what Sapsan can do!
         ### Want to learn more?
@@ -152,6 +147,12 @@ def cnn3d():
         widget_values['backend_list'] = ['Fake', 'MLflow']
         widget_values['backend_selection_index'] = widget_values['backend_list'].index(widget_values['backend_selection'])
         
+    def text_to_list(value):
+        to_clean = ['(', ')', '[', ']', ' ']
+        for i in to_clean: value = value.translate({ord(i) : None})
+        value = list([int(i) for i in value.split(',')])
+        return value
+        
     #show loss vs epoch progress with plotly
     def show_log(progress_slot, epoch_slot):
         from datetime import datetime
@@ -227,10 +228,12 @@ def cnn3d():
                            features=features,
                            target=target,
                            checkpoints=checkpoints,
-                           batch_size=int(widget_values['batch_size']),
-                           checkpoint_data_size=int(widget_values['checkpoint_data_size']),
-                           sampler=sampler)
-        x, y = data_loader.load()
+                           batch_size=text_to_list(widget_values['batch_size']),
+                           input_size=text_to_list(widget_values['input_size']),
+                           sampler=sampler,
+                           shuffle = False,
+                           train_fraction = 1)
+        x, y = data_loader.load_numpy()
         return x, y, data_loader
     
             
@@ -245,13 +248,16 @@ def cnn3d():
         
         #Load the data 
         x, y, data_loader = load_data(widget_values['checkpoints'])
+        y = flatten(y)
+        loaders = data_loader.convert_to_torch([x, y])
+        
         st.write("Dataset loaded...")
         
         #Set the experiment
         training_experiment = Train(backend=tracking_backend,
                                     model=estimator,
-                                    inputs=x, targets=y,
-                                    data_parameters = data_loader.get_parameters(),
+                                    loaders = loaders,
+                                    data_parameters = data_loader,
                                     show_log = False)
         
         #Plot progress        
@@ -272,12 +278,13 @@ def cnn3d():
         #--- Test the model ---
         #Load the test data
         x, y, data_loader = load_data(widget_values['checkpoint_test'])
+        loaders = [x, y]
 
         #Set the test experiment
         evaluation_experiment = Evaluate(backend=tracking_backend,
                                          model=training_experiment.model,
-                                         inputs=x, targets=y,
-                                         data_parameters = data_loader.get_parameters())
+                                         loaders = loaders,
+                                         data_parameters = data_loader)
         
[…]
