+++ sapsan/examples/GUI/st_custom.py
@@ -9,8 +9,7 @@ sys.path.append(str(Path.home())+"/Sapsan/")
 
 from sapsan.lib.backends.fake import FakeBackend
 from sapsan.lib.backends.mlflow import MLflowBackend
-from sapsan.lib.data.hdf5_dataset import HDF5Dataset
-from sapsan.lib.data import EquidistanceSampling
+from sapsan.lib.data import HDF5Dataset, EquidistantSampling, flatten
 from sapsan.lib.estimator import CNN3d, CNN3dConfig
 from sapsan.lib.estimator.cnn.spacial_3d_encoder import CNN3dModel
 from sapsan.lib.experiments.evaluate import Evaluate
@@ -107,6 +106,13 @@ def custom():
         widget_values['backend_list'] = ['Fake', 'MLflow']
         widget_values['backend_selection_index'] = widget_values['backend_list'].index(widget_values['backend_selection'])
         
+    def text_to_list(value):
+        to_clean = ['(', ')', '[', ']', ' ']
+        for i in to_clean: value = value.translate({ord(i) : None})
+        value = list([int(i) for i in value.split(',')])
+        return value
+        
+    #show loss vs epoch progress with plotly
     def show_log(progress_slot, epoch_slot):
         from datetime import datetime
         
@@ -116,7 +122,15 @@ def custom():
         while log_exists == False:
             if os.path.exists(log_path):
                 log_exists = True
+            
             time.sleep(0.1)
+        
+        first_entry = False
+        while first_entry == False:
+            with open(log_path) as file:
+                if len(list(file))>=4: 
+                    first_entry = True
+            time.sleep(0.05)
             
         plot_data = {'epoch':[], 'train_loss':[]}
         last_epoch = 0
@@ -127,31 +141,14 @@ def custom():
             with open(log_path) as file:
                 #get the date of the latest event
                 lines = list(file)
-                latest_time = lines[-4].replace(",",".")
-                latest_time = datetime.strptime(latest_time, '[%Y-%m-%d %H:%M:%S.%f] ')
-
-                #check for the newest entry
-                if start_time > latest_time:
-                    continue
-                else:
-                    current_epoch = int(lines[-2].split('/')[0])
-                    train_loss = float(lines[-2].split('loss=')[-1])
-                    valid_loss = float(lines[-1].split('loss=')[-1])
-
-                '''
-                #to read a .json file
-                data = OrderedDict(json.load(file))
-                elem = list(data.keys())
-
-                if 'epoch' in elem[-1]:
-                    current_epoch = int(elem[-1].rpartition('_')[-1]) + 1
-                else:
-                    current_epoch = -1
-                '''
-            if current_epoch == last_epoch or current_epoch == -1:
+                
+                current_epoch = int(lines[-2].split('/')[0])
+                train_loss = float(lines[-2].split('loss=')[-1])
+                valid_loss = float(lines[-1].split('loss=')[-1])
+                                
+            if current_epoch == last_epoch:
                 pass
             else:     
-                #metrics = data['epoch_%d'%(current_epoch-1)][-1]
                 metrics = {'train_loss':train_loss, 'valid_loss':valid_loss}
                 epoch_slot.markdown('Epoch:$~$**%d** $~~~~~$ Train Loss:$~$**%.4e**'%(current_epoch, metrics['train_loss']))
                 plot_data['epoch'] = np.append(plot_data['epoch'], current_epoch)
@@ -171,49 +168,56 @@ def custom():
                 
                 last_epoch = current_epoch
 
-            if current_epoch == widget_values['n_epochs']:
+            if current_epoch == widget_values['n_epochs']: 
                 return
-            
+                        
             time.sleep(0.1) 
[…]
