+++ sapsan/examples/GUI/st_config_custom.txt
@@ -3,17 +3,17 @@ experiment name = CNN experiment
 mlflow_host = localhost
 mlflow_port = 9000
 edit_port = 8601
-axis = 3
-checkpoint_data_size = 128
-sample_to = 32
-batch_size = 16
+input_size = (32,32,32)
+sample_to = (16,16,16)
+batch_size = (8,8,8)
 n_epochs = 5
 patience = 10
 min_delta = 1e-10
-path = /raid1/pkarpov/CCSN_MRI_SIMS/test/{feature}_s4_it{checkpoint:1.0f}.h5
-checkpoints = 500
-features = vel, dvel, Bvec, dBvec
-target = tn
+path = ../data/t{checkpoint:1.0f}/{feature}_dim32_fm15.h5
+checkpoints = 0
+checkpoint_test = 0
+features = u
+target = u
 backend_selection = Fake
 sampler_selection = Equidistant3D
 
