### 2021-05-25 pikarpov-LANL/Sapsan#99 @kburns
JOSS documentation / paper revisions
Hello, as a part of my review of the Sapsan JOSS paper (https://github.com/openjournals/joss-reviews/issues/3199), here are some questions and comments relating to the documentation and JOSS draft:

- [x] Scope: Overall I think the scope of the software should be specified more clearly, particularly in the web documentation. E.g., is the goal to make a unified interface to models from various ML frameworks that are relevant for learning subgrid closures? The description "Sapsan is a pipeline for easy Machine Learning implementation in scientific projects" is overly broad and the documentation should reflect the current features.

- [x] Need: It's stated that Sapsan makes ML easy, but it would be good to describe some of the specific workflow difficulties that Sapsan helps users avoid, perhaps data manipulation and normalization or hyperparameter selection?

- [x] Installation: It looks like pip installing Sapsan in a clean Python 3.9 environment (on MacOS at least) fails due to an issue with scikit-image wheels for 3.9 (perhaps related to https://github.com/scikit-image/scikit-image/issues/5060). Until this upstream issue is fixed, it should be noted that the pip installation requires Python <= 3.8.

- [x] Examples: Many of the abstraction implementation links on the JOSS-branch README are broken (e.g. "3d dataset" and all the "experiment" links).

- [x] Functionality: The majority of the necessary documentation (functionality and API documentation) is located in the Wiki, but this is not linked from the README page in the JOSS branch.

- [x] Tests: There do not appear to be any automated tests.  A command-line testing interface (e.g. using pytest) that returns a binary pass/fail signal would be preferable to the current suggested test of running one of the notebook examples.

- [x] Community guidelines: There do not appear to be any contribution/support guidelines on the README or Wiki.

- [x] Summary: The summary should be expanded to more clearly describe the scope and functionality of the current codebase.

- [x] State of the field: specific comparisons should be made to existing software packages / workflows.

- [x] Writing: I think the paper should be more specific about how the package can be used. Results from two test problems are shown, but the paper should better indicate what options the user can modify and what general model space is covered by the code, beyond just the two specific setups used in the examples.

- [x] References: the references should be expanded to include the ML packages that Sapsan wraps / relies on (e.g pytorch, scikit, etc.).
