### 2021-06-07 pikarpov-LANL/Sapsan#100 @MilesCranmer
[JOSS Review] recommended revisions
Here is my review for JOSS (https://github.com/openjournals/joss-reviews/issues/3199), apologies for the delay. Here is my current recommended list of revisions. I will post additional messages here with further comments as we go along.

Overall:

- [x] It is a bit unclear who the audience is. Is this: people who do both ML and turbulence? Or, people who do turbulence, but have never tried ML? For those who do both, you may want to present this as a "model zoo" rather than a pipeline, since presumably those who already do ML will have their own pipelines set up and would want to plug in your models, or use your project for baseline modelling. For people who do turbulence but not ML, I think this seems very useful so they don't have to learn the internals of how ML works. But I think, for that audience, there is too much jargon here, and you should consider what changes you can make so that someone completely new to ML would be able to use your project for science.
  - I should mention: for another project (workshop version here: https://simdl.github.io/files/26.pdf, but longer one is in the works) - we ran baselines for different turbulence models, and actually Kim & Alvaro had to implement each one themselves. But it would have been much better if there was a "model zoo" that we could just run automatically, like with Sapsan. I think that would be a really useful area you could satisfy, as more and more people are working on ML & Turbulence.
- [x] README and goals of the project seem much too broad. I think you should focus in on exclusively turbulence modeling problems. Don't aim for generic usage - the feature creep will hurt your project, and prevent you from tuning parts of your software to this specific area which would make it better for your user base. There are far far too many "easy ML pipeline" projects out there, many of them backed by large corporate entities, so I would not try to satisfy that need. Emphasize that your software targets the unsaturated area---ML for turbulence! And really try to make it optimized for that specific problem, in a way that the generic pipelines can't.
  - Take a look through https://github.com/josephmisiti/awesome-machine-learning#python and you can see just how saturated this area has become. I really think it is better to specialize and emphasize your software is better for your particular science, compare to more generic tools.

Code review:

- [x] The software requirements are much too specific. I would change them all to be `>=`, and only if you know of a specific reason to keep it fixed to a version should you change it. Right now when I install it on Google Colab and my base env, I am unable to satisfy version requirements without downgrading my packages.
- [x] Can change git clone -> "pip install git+https://github.com/pikarpov-LANL/Sapsan.git"
  - [x] You should make "setup_gpu.py" part of setup.py. I think to do this, you can set the "extra_requires" part of the setuptools.setup. Then you would set, e.g., "pip install sapsan[gpu]" to get the GPU version.
- [x] However, I think you should not handle all of these requirements. Installing PyTorch can be tricky (look at all the options: http://pytorch.org/), so I would simply direct the user to install torch themselves, and then install your package.
- [x] The directory structure is too hierarchical. Instead of this:
from sapsan.lib.backends.fake import FakeBackend
from sapsan.lib.backends.mlflow import MLflowBackend
from sapsan.lib.data import HDF5Dataset, EquidistantSampling, flatten
from sapsan.lib.estimator import CNN3d, CNN3dConfig
from sapsan.lib.experiments.evaluate import Evaluate
from sapsan.lib.experiments.train import Train
from sapsan.utils.plot import model_graph
```
I would like to use something closer to this:
from sapsan.backends import FakeBackend, MLflowBackend
from sapsan.data import HDF5Dataset, EquidistantSampling, flatten
from sapsan import CNN3d, CNN3dConfig, Evaluate, Train, model_graph
```
You can use the `__init__.py` files to achieve something like this, even with a hierarchical file structure.

Documentation review:

- [x] I think it would be wise to have your example data automatically-downloaded by the Jupyter notebook, or some internal module (just like PyTorch will automatically download MNIST data for you), so one is not forced to clone the repo and work locally.
- [x] For the audience of people who do not normally do ML, I would make a Colab notebook that they can muck around with in-browser, without having to worry about dependency installation. If someone has to set it up on their own computer, I think you might lose half your potential users from the get-go.
  - You can go to this link: http://colab.research.google.com/github/pikarpov-LANL/Sapsan/blob/master/sapsan/examples/cnn_example.ipynb to get the github running. You should then add this link to your README. I would change your package and that notebook until you are able to completely run that notebook in Google colab - it is a good test of your software requirements being flexible enough.

Paper review
- [x] "Thus, it is not surprising that there is considerable skepticism against ML in physical sciences, with astrophysics being no exception."
  - [x] Personally, I don't think the skepticism is from a lack of easy-to-use software. If anything, ML toolkits are probably easier to use than most turbulence simulation software. I think the skepticism *rightly* stems from ML x Physics people not doing enough work to test the generalization of their models to configurations far outside of the test set, where the ML model will often fail, and this has produced skepticism in the community.
- [x] Paper is a bit light on references, and comparison to other work. See https://simdl.github.io/files/26.pdf for a list of other relevant references which you might want to include. Some of these references have code, and some do not.
- [x] Can you justify the use of MLFlow compared to other frameworks? I don't doubt it is a good option, but I am curious to hear why you chose it simply for my own interest and potential use.

Thanks! Let me know if you have any questions about these.
Miles
