### 2021-06-11 7c9a4ef
Generalized requirements + removed PyTorch install
	- no longer need a separate gpu setup
