### 2021-06-11 9f68aa4
Update pypi-release.yml
- Removed GPU release
- Included PyTorch install
