+++ .github/workflows/pypi-release.yml
@@ -25,8 +25,7 @@ jobs:
         --upgrade
         setuptools wheel
     - name: Install PyTorch
-      run: >-
-        python -m
+      run: |
         pip3 install torch==1.8.1+cpu torchvision==0.9.1+cpu torchaudio==0.8.1 -f https://download.pytorch.org/whl/torch_stable.html
     - name: Set version for release
       run: >-
