### 2021-06-15 dce85b0
Updated pytorch module for the latest Catalyst
