+++ tests/examples/test_examples.py
@@ -28,4 +28,7 @@ class TestExamples(unittest.TestCase):
         for nt in self.notebooks():
             _, errors = run_notebook(notebook_path="./runtime_test_resources/"+nt,
                                      resources_path="./")
+			print('-----------------------------')
+			print(errors)
+			print('-----------------------------')
             self.assertEqual(errors, [])
