### 2021-06-22 5f690eb
Fixed log plotting based on Catalyst 21.5
