### 2021-06-24 601483a
requirement updates, hierarchy, minor bugs (#101)
* Generalized requirements + removed PyTorch install

	- no longer need a separate gpu setup

* Update pypi-release.yml

- Removed GPU release
- Included PyTorch install

* Update pythonpackage.yml

* Updated pytorch module for the latest Catalyst

* Updated the example

* Fixing unit test

* Fixed log plotting based on Catalyst 21.5

* Fixed GUI for Catalyst 21.5:wq

* Simplified hierarchy

Co-authored-by: Platon Karpov <plkarpov@ucsc.edu>
