+++ requirements_gpu.txt
@@ -1,23 +1,0 @@
-numpy>=1.19.2
-Click==7.1.2
-torch==1.7.1+cu110
-torchvision==0.8.2+cu110 
-catalyst==20.7
-h5py>=2.10.0
-notebook<6.1.6
-jupyter==1.0.0
-tornado<6.0.0
-matplotlib==3.3.2
-mlflow==1.11.0
-pandas>=1.2.0
-Pillow>=8.1.0
-plotly==4.3.0
-safitty>=1.3
-scikit-image==0.17.2
-scikit-learn==0.23.2
-scipy>=1.5.2
-seaborn==0.11.1
-six==1.15.0
-graphviz==0.14
-streamlit==0.58.0
-opencv-python>=4.5.1
