### 2021-06-25 pikarpov-LANL/Sapsan#100 @pikarpov-LANL
## Overall:

- The target audience is both - people who do turbulence with and without ML. Sapsan is an attempt to structure and standardize ML implementation for turbulence modeling. This will make it much easier to read and implement the models from the papers. In Sapsan, the main focus is on the model structure itself, which is isolated in a separate script without the overhead. Also, I built Sapsan with Docker in mind for straightforward public releases of the models with the publications. In terms of jargon, I agree that I need to simplify the point of entry and I do plan to do so via tutorials.
    - As for the “model zoo”, it would be great! Actually, that’s the direction I want to build Sapsan in. For now, the models that I use and that I find interesting to try for my own research will be published through Sapsan. In addition, I am trying to engage the local LANL community to do the same through Sapsa, but open-sourcing anything at LANL is a big task for the researchers here. On my own, I do not have the capacity to add a significant amount of turbulence models to call Sapsan a “model zoo”; at least not yet.
- This is a very good point. Even though Sapsan is capable of doing pretty much anything related to ML, so are numerous other pipelines. Since Sapsan has been built with turbulence modeling in mind, I brought this focus forward in README and Wiki.

## Code review:

- I generalized the requirements - updated compatibility to Catalyst 21.5 (there have been significant changes)
     - The only thing I didn’t touch is streamlit, which is hard-locked to a specific version. It is fast developing, yet not very well backwards compatible. While GUI is a fun way to demo models for presentation purposes, it won’t be used as much solely for research. Thus, I decided not to spend too much time on its maintenance right now with the latest streamlit versions (though it is fully operational).
- I removed the PyTorch installation and the separated GPU setup since it only different in the CUDA-enabled PyTorch installation. I agree that PyTorch should be installed separately. Wiki [Installation](https://github.com/pikarpov-LANL/Sapsan/wiki/installation) has been updated.
- Hierarchy simplified.

## Documentation review:

- I set up the [google-colab notebook](https://github.com/pikarpov-LANL/Sapsan/wiki/Web), which is identical to the cnn_example.ipynb, except the requirements installation. Now, folks can try out the jupyter notebook and the GUI demo both online.
- data will be updated to a better physical example as right now it is training on low-res velocity and predicting itself. What it will be is the working example from the papers I am writing at the moment - higher resolution JHTDB data to predict a closure turbulence model. The data will be updated once the paper will be in review (in a matter of weeks)

## Paper review:

- I agree with your point - generalization testing is a big roadblock in our field. However, the argument I am trying to make is not around easy-to-use software, but a “black box” nature of ML. The latter is the most often used argument that I hear from the physics community. In my opinion, besides the convoluted nature of classic ML (no pun intended), accessibility is a major factor - not as much to the ML researchers building their own models, but to the physicists reading and evaluating the published work. It is especially relevant to the published work without released code. Through Sapsan I attempt to build a transparent pipeline, which includes physics-based analytic tools, and also bring physics-informed methods to a broader community (which are not trivial to build yourself straight from the papers without the code). While on my own I cannot solve generalizability, I can make improvements on the latter.
- References - no worries, I will send the updated version with more references in the joss github thread for this review, that’s no problem.
- MLflow has been suggested by my collaborator from the industry. Overall, I also considered kubeflow, but MLflow is much easier to track and analyze ensembles of simulations. One can plot various metrics and parameters over hundreds of runs with just a few clicks. Lastly, it is very straightforward to add more parameters to track if one needs to customize further.
