### 2021-06-25 #3199 @pikarpov-LANL
Hi everyone! First and foremost, thank you for your comments. I have made appropriate edits and posted my responses right in the Issues thread.

@MilesCranmer you can check them out [pikarpov-LANL/Sapsan#100](https://github.com/pikarpov-LANL/Sapsan/issues/100)

@kburns I am finishing up the response and will post it here tonight [pikarpov-LANL/Sapsan#99](https://github.com/pikarpov-LANL/Sapsan/issues/99)
