### 2021-06-25 ae03e40
version (#102)
* version update

* Update README.md

Co-authored-by: Platon Karpov <plkarpov@ucsc.edu>
Co-authored-by: pikarpov-LANL <pikarpov-LANL@users.noreply.github.com>
