### 2021-06-25 236ff8d
updates to GUI (#103)
* Updated description in GUI + path fix

Co-authored-by: Platon Karpov <plkarpov@ucsc.edu>
