+++ requirements.txt
@@ -7,7 +7,7 @@ jupyter>=1.0.0
 tornado<6.0.0
 matplotlib>=3.3.2
 mlflow>=1.11.0
-pandas>=1.2.0
+pandas>=1.1.0
 Pillow>=8.1.0
 plotly>=4.3.0
 safitty>=1.3
