### 2021-06-25 7d19b58
up-to-date (#104)
* Heavy updates to GUI interface and website

* minor correction

* minor edit to readme

* Major GUI improvements

	- Ability to edit the model code
	- Launch and see MLflow from GUI
	- Animated training progress with interactive loss plots

* Updated config

* Fixed logging data runaway

	- Checkpoints are not written anymore
	- Re-did the runtime progress plots

* A large number of fixes:

	- GUI now has "custom" section
	- mlflow tracks every re-run of the cell (no need to restart the kernel)
	- mlflow records all the relevant features + plots + model structure
	- backend
[…]
