+++ sapsan/lib/data/hdf5_dataset.py
@@ -2,95 +2,50 @@
 HDF5 dataset classes
 
 Usage:
-    ds = HDF5Dataset(path="/Users/icekhan/Documents/development/myprojects/sapsan/repo/Sapsan/dataset",
-                      features=['u', 'b', 'a',
-                                'du0', 'du1', 'du2',
-                                'db0', 'db1', 'db2',
-                                'da0', 'da1', 'da2'],
-                      target=['tn'],
-                      checkpoints=[0.0, 0.01, 0.025, 0.25])
+    data_loader = HDF5Dataset(path="/path/to/data.h5",
+                      features=['a', 'b'],
+                      target=['c'],
+                      checkpoints=[0.0, 0.01],
+                      batch_size=BATCH_SIZE,
+                      input_size=INPUT_SIZE,
+                      sampler=sampler,
+                      flat = False)
 
-    plugin = HDF5DatasetPyTorchSplitterPlugin(4)
-    loaders = plugin.apply(ds)
+    x, y = data_loader.load_numpy()
 """
 
 from typing import List, Tuple, Dict, Optional
 import numpy as np
 import h5py as h5
-from skimage.util.shape import view_as_blocks
-from sklearn.model_selection import train_test_split
-from torch import from_numpy
-from torch.utils.data import DataLoader, TensorDataset
+import warnings
 
-from sapsan.core.models import Dataset, DatasetPlugin, Sampling
-from sapsan.utils.shapes import split_cube_by_grid, split_square_by_grid
-
-class HDF5DatasetPyTorchSplitterPlugin(DatasetPlugin):
-    def __init__(self,
-                 batch_size: int,
-                 train_size: float = 0.5,
-                 shuffle: bool = True):
-        self.batch_size = batch_size
-        self.train_size = train_size
-        self.shuffle = shuffle
-
-    def apply_on_x_y(self, x, y) -> Dict[str, DataLoader]:
-        x_train, x_test, y_train, y_test = train_test_split(x, y,
-                                                            train_size=self.train_size,
-                                                            shuffle=True)
-
-        train_loader = DataLoader(dataset=TensorDataset(from_numpy(x_train),
-                                                        from_numpy(y_train)),
-                                  batch_size=self.batch_size,
-                                  shuffle=self.shuffle,
-                                  num_workers=4)
-
-        val_loader = DataLoader(dataset=TensorDataset(from_numpy(x_test),
-                                                      from_numpy(y_test)),
-                                batch_size=self.batch_size,
-                                shuffle=self.shuffle,
-                                num_workers=4)
-
-        return {"train": train_loader, "valid": val_loader}
-
-    def apply(self, dataset: Dataset) -> Dict[str, DataLoader]:
-        x, y = dataset.load()
-        return self.apply_on_x_y(x, y)
-
-
-class OutputFlatterDatasetPlugin(DatasetPlugin):
-
-    @staticmethod
-    def _flatten_output(output: np.ndarray):
-        return output.reshape(output.shape[0], -1)
-
-    def apply(self, dataset: Dataset):
-        x, y = dataset.load()
-        return x, self._flatten_output(y)
-
-    def apply_on_x_y(self, x, y):
-        return x, self._flatten_output(y)
-
+from sapsan.core.models import Dataset, Sampling
+from sapsan.utils.shapes import split_cube_by_batch, split_square_by_batch
+from .data_functions import torch_splitter, flatten
 
 class HDF5Dataset(Dataset):
     def __init__(self,
                  path: str,
-                 features: List[str],
-                 target: List[str],
-                 checkpoints: List[int],
-                 grid_size: int = 64,
-                 checkpoint_data_size: int = 128,
+                 input_size,
+                 checkpoints: List[int] = [0],
[…]
