+++ sapsan/lib/experiments/__init__.py
@@ -1,5 +1,2 @@
-from .evaluate_3d import Evaluate3d
-from .evaluate_autoencoder import EvaluateAutoencoder
-from .evaluate_flatten import EvaluateFlatten
+from .evaluate import Evaluate
 from .train import Train
-
