+++ sapsan/lib/backends/mlflow.py
@@ -1,18 +1,38 @@
 import mlflow
+from threading import Thread
+import os
+import time
 
 from sapsan.core.models import ExperimentBackend
 
 
 class MLflowBackend(ExperimentBackend):
-    def __init__(self, name: str, host: str, port: int):
+    def __init__(self, name: str = 'experiment', host: str = 'localhost', port: int = 9000):
         super().__init__(name)
         self.host = host
         self.port = port
+        
         self.mlflow_url = "http://{host}:{port}".format(host=host,
                                                         port=port)
         mlflow.set_tracking_uri(self.mlflow_url)
-        self.experiment_id = mlflow.set_experiment(name)
-
+        try:
+            self.experiment_id = mlflow.set_experiment(name)
+            print("mlflow ui is already running at %s:%s"%(self.host, self.port))
+        except: 
+            print("starting mlflow ui, please wait ...")
+            self.start_ui()
+            self.experiment_id = mlflow.set_experiment(name)
+            print("mlflow ui is running at %s:%s"%(self.host, self.port))
+    
+    def start_ui(self):
+        mlflow_thread = Thread(target=
+                       os.system("mlflow ui --host %s --port %s &"%(self.host, self.port)))
+        mlflow_thread.start()
+        time.sleep(5)
+        
+    def start(self, run_name: str):
+        mlflow.start_run(run_name = run_name)
+        
     def log_metric(self, name: str, value: float):
         mlflow.log_metric(name, value)
 
@@ -21,3 +41,6 @@ class MLflowBackend(ExperimentBackend):
 
     def log_artifact(self, path: str):
         mlflow.log_artifact(path)
+
+    def end(self):
+        mlflow.end_run()

[…]
