### 2021-07-01 5542da9
Simplified Docker (minor)
