### 2021-07-01 pikarpov-LANL/Sapsan#99 @pikarpov-LANL
## Documentation
#### Scope:
- I agree that I should have a tighter focus, which Sapsan already has. There is no reason to make it general. I edited the README accordingly.

#### Need:
- README updated
#### Installation:
- I had !=3.9 statement in the setup.py + a pip clarification of python = 3.7 and 3.8. Thus, it is strange that Sapsan was installed fine in python=3.9. Nonetheless, I put in an explicit statement of <=3.8
#### Examples:
- JOSS branch has not been merged for a while as I only keep the paper there. In either case - updated.
#### Functionality:
- Link to the Wiki is now front and center!
#### Tests:
- There has always been a few unit tests for data loading and unit tests with all current ML models included in Sapsan through jupyter notebooks via pytest. Is there anything else I should add?
#### Community guidelines:
- I have added a section describing how to contribute to Sapsan. Please review [Community Guidelines](https://github.com/pikarpov-LANL/Sapsan/wiki/Community-Guidelines).
