+++ docs/joss/paper.md
@@ -96,6 +96,17 @@ For further information on each stage please refer to [Sapsan's Wiki on Gihub](h
 
 ![High-level overview of ``Sapsan's`` workflow.](Sapsan_highlevel_overview.png)
 
+### Dependencies
+Here only the core dependencies will be covered. Please refer to [GitHub](https://github.com/pikarpov-LANL/Sapsan) for the full list.
+
+* __Training__ 
+   * __PyTorch:__ `Sapsan`, at large, relies on PyTorch to configure and train ML models. Thus, the parameters in the aformentioned __Model Set Up__ stage should be configured with PyTorch functions. [Convolutional Neural Network (CNN)](https://github.com/pikarpov-LANL/Sapsan/blob/master/sapsan/examples/cnn_example.ipynb) and [Physics-Informed Convolutional Auto Encoder (PICAE)](https://github.com/pikarpov-LANL/Sapsan/blob/master/sapsan/examples/picae_example.ipynb) examples included with `Sapsan` are based on PyTorch.
+      * __Catalyst:__ used as part of the backend to configure early-stopping of the model and logging
+   * __scikit-learn:__ the framework supports it, as shown in the [Kernel Ridge Regression (KRR)](https://github.com/pikarpov-LANL/Sapsan/blob/master/sapsan/examples/krr_example.ipynb) example in `Sapsan`. However, due to lack of scalability and features, it is advised to use PyTorch based setup.
+* __Tracking__
+   * __MLflow:__ allows for tracking across large quantities of training and evaluation runs through an intuitive web interface. Beyond the few default parameters, a user can include custom parameters to be tracked. 
+
+
 
 # Applications
 
