+++ docs/joss/paper.bib
@@ -73,6 +73,13 @@ archivePrefix = {arXiv},
   howpublished = {\url{https://github.com/streamlit/streamlit}},
 }
 
+@software{click,
+  author = {Armin Ronacher},
+  title = {Click},
+  url = {https://click.palletsprojects.com/},
+  year = 2021
+}
+
 @misc{mlflow_github,
   author = {Databricks, Inc},
   title = {MLflow},
