### 2021-07-02 pikarpov-LANL/Sapsan#99 @pikarpov-LANL
## Software paper

#### Summary:

- I included more details/examples of Sapsan’s capabilities in the Summary. Please let me know if this is enough or if I should expand further.

#### State of the field:

- As far as I am concerned, there is nothing like Sapsan bringing ML and turbulence modeling. I have tightened the focus on turbulence modeling in the paper, instead of the general applications in physical sciences.

#### Writing:

- I added a bit more text to the Framework section to indicate what the user can adjust in Sapsan. Please let me know if this is sufficient or if I should include a separate section.

#### References:

- Added references to the core packages under Dependencies in section Framework of the paper.
