### 2021-07-13 471448b
Slight fix to examples + turned off parallel GPU
