### 2021-07-13 45ac31e
Parallel GPU halt, Docker, minor fixes (#105)
* Updated description in GUI + path fix

* Simplified Docker (minor)

* Slight fix to examples + turned off parallel GPU

Co-authored-by: Platon Karpov <plkarpov@ucsc.edu>
Co-authored-by: pikarpov-LANL <pikarpov-LANL@users.noreply.github.com>
