+++ Dockerfile
@@ -1,6 +1,6 @@
 FROM python:3.8.5-slim
 
-# remember to expose the port your app'll be exposed on.
+# remember to expose the port your app will run on
 EXPOSE 7654
 
 ENV GIT_PYTHON_REFRESH=quiet
@@ -8,10 +8,8 @@ RUN pip install -U pip
 
 RUN pip install sapsan
 
-# copy into a directory of its own (so it isn't in the toplevel dir)
+# copy into a directory of its own (so it isn't in the top-level dir)
 COPY sapsan/examples/cnn_example.ipynb sapsan_docker_examples/
-COPY sapsan/examples/krr_example.ipynb sapsan_docker_examples/
-COPY sapsan/examples/GUI sapsan_docker_examples/GUI/
 COPY sapsan/examples/data sapsan_docker_examples/data/
 WORKDIR /sapsan_docker_examples
 
