### 2021-07-13 ae7a72a
Fixed incorrect dim of loaded data if batch_num==1
