### 2021-07-13 3f67af3
Fixed loading of muti-feature x and y simultaniously
