### 2021-07-13 6d569e1
Corrected n_output_channels
	- if only a single feature is chosen to test
