### 2021-07-14 0aaf3a6
Fix: MLflow metrics logging, DDP functionality
	- multi-gpu training via Distributed Data Parallel (DDP)
	  (needs to specified in model config)
	- fixes in train and eval to account for DDP
	- Examples: comments clarified, simplified mlflow
