### 2021-07-14 2f06621
Added self.kwargs in estimator config
	- estimator template as well
	- allows to turn on ddp + other pars in the future
