+++ sapsan/lib/estimator/picae/picae_encoder.py
@@ -213,6 +213,7 @@ class PICAEConfig(EstimatorConfig):
         self.logdir = logdir
         self.patience = patience
         self.min_delta = min_delta
+        self.kwargs = kwargs
         
         #a few custom names for parameters to record in mlflow
         self.parameters = {
