### 2021-07-15 887769e
Multi-GPU support (DDP), MLflow fixes, Gradient and Dynamic Smagorisnky models (#106)
* Fixed incorrect dim of loaded data if batch_num==1

* Fixed loading of muti-feature x and y simultaniously

* Corrected n_output_channels

	- if only a single feature is chosen to test

* Fix: MLflow metrics logging, DDP functionality

	- multi-gpu training via Distributed Data Parallel (DDP)
	  (needs to specified in model config)
	- fixes in train and eval to account for DDP
	- Examples: comments clarified, simplified mlflow

* Added self.kwargs in estimator config

	- estimator template as well
	- allows
[…]
