+++ sapsan/lib/estimator/cnn/spacial_3d_encoder.py
@@ -15,16 +15,15 @@ from sapsan.lib.estimator.pytorch_estimator import TorchEstimator
 from sapsan.lib.data import get_loader_shape
 
 
-class CNN3dModel(torch.nn.Module):
+class CNN3dModel(torch.nn.ModuleDict):
     def __init__(self, D_in = 1, D_out = 1):
         super(CNN3dModel, self).__init__()
         
         self.conv3d = torch.nn.Conv3d(D_in, D_in*2, kernel_size=2, stride=2, padding=1)
-        self.pool = torch.nn.MaxPool3d(kernel_size=2, padding=1)
         self.conv3d2 = torch.nn.Conv3d(D_in*2, D_in*2, kernel_size=2, stride=2, padding=1)
-        self.pool2 = torch.nn.MaxPool3d(kernel_size=2, padding=1)
         self.conv3d3 = torch.nn.Conv3d(D_in*2, D_in*4, kernel_size=2, stride=2, padding=1)
-        self.pool3 = torch.nn.MaxPool3d(kernel_size=2)
+        self.pool = torch.nn.MaxPool3d(kernel_size=2, padding=1)
+        self.pool2 = torch.nn.MaxPool3d(kernel_size=2)
 
         self.relu = torch.nn.ReLU()
 
@@ -37,9 +36,9 @@ class CNN3dModel(torch.nn.Module):
         c1 = self.conv3d(x)
         p1 = self.pool(c1)
         c2 = self.conv3d2(self.relu(p1))
-        p2 = self.pool2(c2)
+        p2 = self.pool(c2)
         c3 = self.conv3d3(self.relu(p2))
-        p3 = self.pool3(c3)
+        p3 = self.pool2(c3)
 
         v1 = p3.view(p3.size(0), -1)
 
@@ -47,7 +46,7 @@ class CNN3dModel(torch.nn.Module):
         l2 = self.linear2(l1)   
 
         return l2
-    
+
     
 class CNN3dConfig(EstimatorConfig):
     def __init__(self,
@@ -60,12 +59,13 @@ class CNN3dConfig(EstimatorConfig):
         self.logdir = logdir
         self.patience = patience
         self.min_delta = min_delta
+        self.kwargs = kwargs
         self.parameters = {
                         "model - n_epochs": self.n_epochs,
                         "model - min_delta": self.min_delta,
                         "model - patience": self.patience,
                     }
-
+        
     @classmethod
     def load(cls, path: str):
         with open(path, 'r') as f:
@@ -90,7 +90,7 @@ class CNN3d(TorchEstimator):
         optimizer = torch.optim.Adam(model.parameters(), lr=0.001)
         loss_func = torch.nn.MSELoss()  # torch.nn.SmoothL1Loss()
         scheduler = torch.optim.lr_scheduler.ReduceLROnPlateau(optimizer,
-                                                               patience=3,
+                                                               patience=self.config.patience,
                                                                min_lr=1e-5) 
         
         trained_model = self.torch_train(loaders, model, 

[…]
