### 2021-07-16 6fb0cbb
Minor edits to PowerSpectrum method
	-proper return of k and E(k)
