### 2021-07-16 8ea7176
A few assertion checks added to physics.py methods
