### 2021-07-16 560aba3
Physics methods: added assertion checks, comments, quality of life (#107)
* Minor edits to PowerSpectrum method

	-proper return of k and E(k)

* A few assertion checks added to physics.py methods

Co-authored-by: Platon Karpov <plkarpov@ucsc.edu>
Co-authored-by: pikarpov-LANL <pikarpov-LANL@users.noreply.github.com>
