### 2021-07-17 #3199 @dfm
@MilesCranmer, @kburns, @pikarpov-LANL: I wanted to check in here to see where we stand. It looks to me like we're currently waiting on @MilesCranmer and @kburns to take a look at @pikarpov-LANL's responses on the review issues (https://github.com/pikarpov-LANL/Sapsan/issues/99, https://github.com/pikarpov-LANL/Sapsan/issues/100). Is that correct and, if so, @MilesCranmer, @kburns can you make sure that you keep this on your radar and let me know if you have issues or questions? Thanks all!!
