### 2021-07-17 #3199 @pikarpov-LANL
@dfm indeed! I have responded to the comments and have fixed up all the noted issues in the code.
