### 2021-07-21 c7e594a
Default FakeBackend if not specified & Simplified Samping
	- no longer need to pass original shape
	- it is determined from the passed data
