### 2021-07-21 15c8fdd
Tensor, Sampling, Backend, Examples (#108)
* Default FakeBackend if not specified & Simplified Samping

	- no longer need to pass original shape
	- it is determined from the passed data

* Fixed Tensor calculation indices 
* Simplified examples according to the changes above

Co-authored-by: Platon Karpov <plkarpov@ucsc.edu>
Co-authored-by: pikarpov-LANL <pikarpov-LANL@users.noreply.github.com>
