### 2021-07-23 d51024e
Fixed input data size processing without sampling
