### 2021-07-23 c7af1de
Added "only_x_components" flag for tensor calc
	- at times full tensor is too taxing on memory
