### 2021-07-23 3c53c9a
Added a warning on dealing with irregular shapes
	- the funciton will try to match the requested
	  dimension, but if it can't - a warning is issued
