+++ sapsan/lib/data/hdf5_dataset.py
@@ -140,12 +140,16 @@ class HDF5Dataset(Dataset):
             print("Loading '%s' from file '%s'"%(key, self._get_path(checkpoint, columns[col])))
             
             data = file.get(key)
+            print(data.shape)
+            if (self.axis==3 and len(np.shape(data))==5) or (self.axis==2 and len(np.shape(data))==4):
+                print("Warning: combining axis for %s"%key)
+                data = np.reshape(data, np.append(data.shape[0]*data.shape[1],data.shape[2:]))
             
             if (self.axis==3 and len(np.shape(data))==3) or (self.axis==2 and len(np.shape(data))==2): 
                 data = [data]     
             all_data.append(data)
             
-        # input_data shape ex: (features, 128, 128, 128)        
+        # input_data shape ex: (features, 128, 128, 128) 
         input_data = np.vstack(all_data)
 
         # downsample if needed
@@ -165,25 +169,24 @@ class HDF5Dataset(Dataset):
 
 
     def _load_data_numpy(self) -> Tuple[np.ndarray, np.ndarray]:
-        x = list()
-        y = list()
         for checkpoint in self.checkpoints:
             features_checkpoint_batch = self._get_input_data(checkpoint, 
                                                              self.features, self.features_label)
-            x.append(features_checkpoint_batch)
-            x = np.vstack(x)
+            
+            if checkpoint == self.checkpoints[0]: x = features_checkpoint_batch
+            else: x = np.vstack((x, features_checkpoint_batch))
+                
             print('Loaded INPUT data shape', x.shape)
                         
             if self.target!=None:
                 target_checkpoint_batch = self._get_input_data(checkpoint, 
                                                                self.target, self.target_label)
-                y.append(target_checkpoint_batch)
-                y = np.vstack(y)                        
+                if checkpoint == self.checkpoints[0]: y = features_checkpoint_batch
+                else: y = np.vstack((y,features_checkpoint_batch))                       
                 print('Loaded TARGET data shape', y.shape)
                 
-                return x, y
-                
-            else: return x
+        if self.target!=None: return x, y
+        else: return x
     
     
     def _check_batch_size(self):
