### 2021-07-23 2356938
Fixed split by batches + default batch_num=1
	- hence batch_size = input_size if it is not specified
