### 2021-07-23 4188ba5
Various Pipeline Fixes (#109)
* Default FakeBackend if not specified & Simplified Sampling

	- no longer need to pass original shape
	- it is determined from the passed data

* Fixed input data size processing without sampling

* Added "only_x_components" flag for tensor calc

	- at times full tensor is too taxing on memory

* Added a warning on dealing with irregular shapes

	- the function will try to match the requested
	  dimension, but if it can't - a warning is issued

* Fixed multi-checkpoint loading for hdf5 module

* Fixed split by batches + default batch_num=1

	- hence batch_size = 
[…]
