### 2021-07-25 pikarpov-LANL/Sapsan#99 @kburns
Thanks for addressing all of these points!  I'm trying to install the latest version now in a conda environment with python 3.8.10, but I'm getting the error:

`ERROR: Package 'sapsan' requires a different Python: 3.8.10 not in '<=3.8,>=3.7'`

I think the specifier should be changed to `<3.9` instead of `<=3.8` so that all `3.8.*` releases are included.
