### 2021-07-25 pikarpov-LANL/Sapsan#99 @pikarpov-LANL
I fixed up the requirements issue and released a new version with various bug fixes - please make sure you install the latest version (v0.2.10).

Good note on testing! I will add it through Click interface and work on the manuscript edits you mentioned later tonight.
