### 2021-07-27 659252d
Fixed notebook testing, mlflow tracking, examples, PICAE device
	- Notebook testing errors trigger pytest
	- For evaluation, some model training pars
	  and metrics are also recorded to make it easier
	  to see with what model the prediction was done
	- Minor formatting fixed in examples
	- removed redundant torch.device pass for PICAE estimator
