### 2021-07-28 2ac7ae5
Simplified Estimator Template + Fixed save and load model config
	- ModelConfig only has __init__
	- moved load() to backend, removed to_dict()
	- Save config now saves all input parameters
	  instead of only the ones tracked by MLflow
	- Loading is now consistent between PyTorch
	  and scikit-learn based models
	- changes reflected in the estimator template + wiki
	- added load_config() under core Estimator class in models.py
	  to enforce it to be included in custom estimators
	  (most cases won't be affected, as it is part of the backend
	  pytorch_estimator.py)
