+++ sapsan/lib/estimator/picae/picae_encoder.py
@@ -214,23 +214,13 @@ class PICAEConfig(EstimatorConfig):
         self.min_delta = min_delta
         self.kwargs = kwargs
         
-        #a few custom names for parameters to record in mlflow
+        #everything in self.parameters will get recorded by MLflow
         self.parameters = {
                         "model - n_epochs": self.n_epochs,
                         "model - min_delta": self.min_delta,
                         "model - patience": self.patience,
                     }
-
-    @classmethod
-    def load(cls, path: str):
-        with open(path, 'r') as f:
-            cfg = json.load(f)
-            return cls(**cfg)
-
-    def to_dict(self):
-        return self.parameters    
-    
-    
+        
     
 class PICAE(TorchEstimator):
     def __init__(self, config = PICAEConfig(), 
