### 2021-07-28 1473a2e
Minor fixed to cnn encoder
