### 2021-07-28 1fe215d
Fixed set_device for PICAE
