### 2021-07-28 27e4fb2
Notebook testing, Device in PICAE, Simplified Estimator Config() (#111)
* Fixed notebook testing, mlflow tracking, examples, PICAE device

	- Notebook testing errors trigger pytest
	- For evaluation, some model training pars
	  and metrics are also recorded to make it easier
	  to see with what model the prediction was done
	- Minor formatting fixed in examples
	- removed redundant torch.device pass for PICAE estimator

* Simplified Estimator Template + Fixed save and load model config

	- ModelConfig only has __init__
	- moved load() to backend, removed to_dict()
	- Save config now saves all 
[…]
