### 2021-07-28 d830dbf
version + minor (#112)
* Update setup.py

* version update

Co-authored-by: Platon Karpov <plkarpov@ucsc.edu>
Co-authored-by: pikarpov-LANL <pikarpov-LANL@users.noreply.github.com>
