### 2021-08-10 b3a2477
Updated version logging - provides metadata
	- sapsan.__version__ now prints correctly
