+++ setup.py
@@ -1,17 +1,15 @@
 import setuptools
+from sapsan._version import __version__
 
 with open("README.md", "r") as fh:
     long_description = fh.read()
-
-with open("version", "r") as f:
-    version = f.read()
 
 with open("requirements.txt", 'r') as f:
     requirements = [i.strip() for i in f.readlines()]
     
 setuptools.setup(
     name="sapsan",
-    version=version,
+    version=__version__,
     author="Platon Karpov, Iskandar Sitdikov",
     author_email="plkarpov@ucsc.edu",
     description="Sapsan project",
