+++ sapsan/_version.py
@@ -1,0 +1,1 @@
+__version__ = 'v0.2.11'
