### 2021-08-11 7abfe1a
Major overhaul of the CLI
	- Significantly simplified `sapsan create -n {name}`
	- Cleaned up the template structure
	- Much easier to navigate and get started
	- Pre-filled all templates with the custom project name
	- Creates the Jupyter Notebook as well to get started
	- Updated CLI options nad --help
