+++ sapsan/core/cli/cli.py
@@ -1,12 +1,18 @@
 import os
 import click
+import jupytext
+import nbformat
 
+from sapsan._version import __version__
+
+from sapsan.core.cli.templates.notebook import get_template as get_notebook_template
 from sapsan.core.cli.templates.estimator import get_template as get_estimator_template
+from sapsan.core.cli.templates.docker import get_template as get_dockerfile_template
+from sapsan.core.cli.templates.makefile import get_template as get_makefile_template
 from sapsan.core.cli.templates.dataset import get_template as get_dataset_template
 from sapsan.core.cli.templates.experiment import get_template as get_experiment_template
 from sapsan.core.cli.templates.runner import get_template as get_runner_template
 from sapsan.core.cli.templates.readme import get_readme_template
-from sapsan.core.cli.templates.docker import get_dockerfile_template
 from sapsan.core.cli.templates.setup import get_setup_template
 from sapsan.core.cli.templates.test import get_template as get_tests_template
 from sapsan.core.cli.templates.actions import (TEST_TEMPLATE, PYPI_TEMPLATE,
@@ -17,8 +23,36 @@ def create_init(path: str):
     with open("{path}/__init__.py".format(path=path), "w") as file:
         file.write("")
 
-
 def setup_project(name: str):
+    click.echo("Created...")
+    os.mkdir(name)
+    click.echo("Project Folder:             {name}/".format(name=name))
+
+    os.mkdir("./{name}/data".format(name=name))
+    click.echo("Data Folder:                {name}/data/".format(name=name))
+
+    with open("./{name}/{name}_estimator.py".format(name=name), "w") as file:
+        file.write(get_estimator_template(name))
+        click.echo("Estimator Template:         {name}/{name}_estimator.py".format(name=name))
+        
+    with open("./{name}/{name}.py".format(name=name), "w") as file:
+        file.write(get_notebook_template(name))
+
+    ntbk = jupytext.read("./{name}/{name}.py".format(name=name)) 
+    nbformat.write(ntbk, "./{name}/{name}.ipynb".format(name=name))
+    os.remove("./{name}/{name}.py".format(name=name))
+    click.echo("Jupyter Notebook Template:  {name}/{name}.ipynb".format(name=name))
+    
+    with open("./{name}/Dockerfile".format(name=name), "w") as file:
+        file.write(get_dockerfile_template(name))
+        click.echo("Docker Template:            {name}/Dockerfile".format(name=name))
+        
+    with open("./{name}/Makefile".format(name=name), "w") as file:
+        file.write(get_makefile_template(name))
+        click.echo("Docker Makefile:            {name}/Makefile".format(name=name))
+    
+
+def setup_package(name: str):
     os.mkdir(name)
     os.mkdir('./{name}/.github'.format(name=name))
     os.mkdir('./{name}/.github/workflows'.format(name=name))
@@ -36,13 +70,21 @@ def setup_project(name: str):
     create_init("./{name}/{name}/experiment".format(name=name))
     click.echo("Marked folders as packages.")
 
-    with open("./{name}/version".format(name=name), "w") as file:
-        file.write("0.0.1")
+    with open("./{name}/{name}/_version.py".format(name=name), "w") as file:
+        file.write("__version__ = v0.0.1")
         click.echo("Created version file.")
 
-    with open("./{name}/{name}/estimator/{name}_estimator.py".format(name=name), "w") as file:
+    with open("./{name}/{name}_estimator.py".format(name=name), "w") as file:
         file.write(get_estimator_template(name))
         click.echo("Created estimator file.")
+        
+    with open("./{name}/{name}.py".format(name=name), "w") as file:
+        file.write(get_notebook_template(name))
+
+    ntbk = jupytext.read("./{name}/{name}.py".format(name=name)) 
[…]
