+++ requirements.txt
@@ -1,4 +1,4 @@
-numpy>=1.19.2
+numpy>=v1.19.2
 Click>=7.1.2
 catalyst>=21.5
 h5py>=2.10.0
@@ -19,3 +19,4 @@ six>=1.15.0
 graphviz>=0.14
 streamlit==0.58.0
 opencv-python>=4.5.1
+jupytext>=1.11
