### 2021-08-11 322352c
Added pytest to the list of requirements
