+++ requirements.txt
@@ -20,3 +20,4 @@ graphviz>=0.14
 streamlit==0.58.0
 opencv-python>=4.5.1
 jupytext>=1.11
+pytest>=6.2
