+++ .github/workflows/pythonpackage.yml
@@ -26,5 +26,4 @@ jobs:
         pip install -r requirements.txt
     - name: Test with pytest
       run: |
-        pip install pytest
-        pytest
+        pytest sapsan/tests/
