### 2021-08-11 714d7ad
Added `sapsan test` command + simplified tests interface
	- moved tests to sapsan, so it is always accessible
	  when installed via pip
	- notebook tests don't create a separate folder, but
	  test on the existing exAmple notebooks
	- `sapsan test` allows for the user to run pytest tests
	  to check if everything was installed correctly
