### 2021-08-11 8be3e1c
CLI `sapsan get_examples`: copies examples into ./sapsan_examples
