### 2021-08-12 be06c45
CLI Overhaul (#113)
* updated _version.py logging

* Updated version logging - provides metadata

	- sapsan.__version__ now prints correctly

* Major overhaul of the CLI

	- Significantly simplified `sapsan create -n {name}`
	- Cleaned up the template structure
	- Much easier to navigate and get started
	- Pre-filled all templates with the custom project name
	- Creates the Jupyter Notebook as well to get started
	- Updated CLI options nad --help

* Added pytest to the list of requirements

* changed pytest tests location

* Added `sapsan test` command + simplified tests interface

	- moved te
[…]
