+++ .github/workflows/pypi-release.yml
@@ -27,6 +27,9 @@ jobs:
     - name: Install PyTorch
       run: |
         pip3 install torch==1.8.1+cpu torchvision==0.9.1+cpu torchaudio==0.8.1 -f https://download.pytorch.org/whl/torch_stable.html
+    - name: Install dependencies
+      run: |        
+        pip3 install -r requirements.txt
     - name: Set version for release
       run: >-
         rm sapsan/_version.py &&
