+++ docs/joss/paper.md
@@ -31,57 +31,38 @@ bibliography: paper.bib
 ---
 
 # Summary
-[Sapsan](https://github.com/pikarpov-LANL/Sapsan) is a framework to make Machine Learning (ML) more accessible in the study of turbulence, with a focus on astrophysical applications. Sapsan includes modules to load, filter, sample down, batch, and split the data from hydrodynamic (HD) simulations for training and validation. Next, the framework includes built-in conventional and physics-informed published estimators that were used for turbulence modeling. This ties into Sapsan's custom estimator module, i.e. ability to focus on a custom model design layer-by-layer, which is the core benefit of using the framework. To share your custom model, every new project created via Sapsan comes with pre-filled, ready for release Docker files. Furthermore, training and evaluation modules are included with Sapsan. The latter, among other features, includes construction of power spectra and comparison to established analytical turbulence closure models, such as a gradient model. Thus, Sapsan takes out all the hard work from data preparation and analysis, leaving one focused on the ML model design itself.
+[Sapsan](https://github.com/pikarpov-LANL/Sapsan) is a framework to make Machine Learning (ML) more accessible in the study of turbulence, focusing on astrophysical applications. Sapsan includes modules to load, filter, subsample, batch, and split the data from hydrodynamic (HD) simulations for training and validation. Next, the framework includes built-in conventional and physics-informed published estimators that were used for turbulence modeling. This ties into Sapsan's custom estimator module, aimed at designing a custom ML model layer-by-layer, which is the core benefit of using the framework. To share your custom model, every new project created via Sapsan comes with pre-filled, ready-for-release Docker files. Furthermore, training and evaluation modules come with Sapsan as well. The latter, among other features, includes the construction of power spectra and comparison to established analytical turbulence closure models, such as a gradient model. Thus, Sapsan takes out all the hard work from data preparation and analysis, leaving one focused on the ML model design itself.
 
 # Statement of Need
 
-It has been challenging for domain sciences to adapt Machine Learning (ML) for their respective projects. 
-From the authors' background, it has been particularly difficult to employ ML for physical simulations, modeling turbulence. 
-It is rather challenging to prove that an ML model has arbitrarily learned the laws of physics embedded into the problem 
-with the ability to extrapolate within the parameter-space of the simulation. Inability to directly infer the prediction 
-capabilities of ML is one of the major causes behind the slow adaption rates, however, the effectiveness of ML cannot be ignored by the community.
+It has been challenging for domain sciences to adopt Machine Learning (ML) for their respective projects, particularly for physical simulations modeling turbulence. It is rather challenging to prove that an ML model has arbitrarily learned the laws of physics embedded into the problem with the ability to extrapolate within the parameter-space of the simulation. The inability to directly infer the prediction capabilities of ML is one of the major causes behind the slow adaption rates; however, the community cannot ignore the effectiveness of ML.
 
-Turbulence is ubiquitous in astrophysical environments; however, it involves physics at a vast range of temporal and spatial 
-scales, making accurate fully-resolved modeling difficult. Various analytical turbulence models have been developed to be used 
-in simulations using time or spatial averaged governing equations, such as in RANS (Reynolds-averaged Navier-Stokes) and LES (Large Eddy Simulation), 
-but accuracy is lacking. In search of better methods to model 
-turbulence in core-collapse supernovae, it became apparent that ML has great potential 
[…]
