### 2021-08-12 pikarpov-LANL/Sapsan#99 @pikarpov-LANL
Please check out the latest v0.3.0 - I have made significant changes to the CLI, further simplifying getting started on a custom project/estimator + added the testing routines to the CLI. Manuscript changes per your suggestions have been made as well.
