+++ README.md
@@ -1,6 +1,6 @@
 # Sapsan  <a href="http://sapsan.app"><img src="https://github.com/pikarpov-LANL/Sapsan/blob/images/docs/images/logo3_black_slim_notitle_whitebg.png?raw=true"  alt="Sapsan logo" align="right" width="100"></a>
 
-Sapsan is a pipeline for Machine Learning (ML) based turbulence modeling. While turbulence is important in a wide range of mediums, the pipeline primarily focuses on astrophysical application. With Sapsan, one can create their own custom models or use either conventional or physics-informed ML approaches for turbulence modeling included with the pipeline ([estimators](https://github.com/pikarpov-LANL/Sapsan/wiki/Estimators)). For example, Sapsan features ML models in its set of tools to accurately capture the turbulent nature applicable to Core-Collapse Supernovae.
+Sapsan is a pipeline for Machine Learning (ML) based turbulence modeling. While turbulence is important in a wide range of mediums, the pipeline primarily focuses on astrophysical application. With Sapsan, one can create their own custom models or use either conventional or physics-informed ML approaches for turbulence modeling included with the pipeline ([estimators](https://github.com/pikarpov-LANL/Sapsan/wiki/Estimators)). Sapsan is designed to take out all the hard work from data preparation and analysis, leaving you focused on ML model design, layer by layer.
 
 Feel free to check out a website version at [sapsan.app](http://sapsan.app). The interface is indentical to the GUI of the local version of Sapsan, except lacking the ability to edit the model code on the fly.
 
@@ -10,37 +10,44 @@ Please refer to Sapsan's github wiki to learn more about framework's details and
 
 ## Quick Start
 
-#### 1. Install PyTorch (prerequisite)
+### 1. Install PyTorch (prerequisite)
 Sapsan can be run on both cpu and gpu. Please follow the instructions on [PyTorch](https://pytorch.org/get-started/locally/) to install the latest version (torch>=1.7.1 & CUDA>=11.0).
 
-#### 2. Clone from git (recommended)
-```shell script
+### 2. Install via pip (recommended)
+```
+pip install sapsan
+```
+
+#### OR Clone from git
+```
 git clone https://github.com/pikarpov-LANL/Sapsan.git
 cd Sapsan/
 python setup.py install
 ```
 
-#### OR Install via pip
-```shell script
-pip install sapsan
-```
-
 Note: see [Installation Page](https://github.com/pikarpov-LANL/Sapsan/wiki/Installation/) on the Wiki for complete instructions with Graphviz and Docker installation.
 
-#### Run Examples
+### 3. Test Installation
 
-To make sure everything is alright and to familiarize yourself with the interface, please run the following CNN example on 3D data:
-```shell script
-jupyter notebook sapsan/examples/cnn_example.ipynb
+To make sure everything is alright, run a test of your setup:
 ```
-alternatively, you can try out the physics-informed convolutional auto-encoder (PICAE) example on random 3D data:
-```shell script
-jupyter notebook sapsan/examples/picae_example.ipynb
+sapsan test
 ```
-or a KRR example on 2D data:
-```shell script
-jupyter notebook sapsan/examples/krr_example.ipynb
+
+### 4. Run Examples
+
[…]
