### 2021-08-13 f70891d
Added `device` option to config
	- sets a specific device to run the models on,
	  either cpu or cuda
