### 2021-08-18 ce09ee3
Learning rate parameters are no longer hard-coded
	- also all 'self' vars in ModelConfig get recorded by default
