### 2021-08-18 0b6a316
Estimator Template + Config updates (#119)
* Added `device` option to config
	- sets a specific device to run the models on,
	  either cpu or cuda

* Learning rate parameters are no longer hard-coded

	- also all 'self' vars in ModelConfig get recorded by default

Co-authored-by: Platon Karpov <plkarpov@ucsc.edu>
Co-authored-by: pikarpov-LANL <pikarpov-LANL@users.noreply.github.com>
