### 2021-08-21 7bd91f5
Fixed save/load models, added sklearn_backend, torch_backend
	- sklearn estimator has a structure close to torch
	- pytorch_estimator -> torch_backend
	- cleared up variable name conventions throughout
