+++ sapsan/lib/estimator/picae/picae_encoder.py
@@ -13,7 +13,7 @@ from torch.utils import data
 from torch.utils.data import DataLoader
 
 from sapsan.core.models import EstimatorConfig
-from sapsan.lib.estimator.pytorch_estimator import TorchEstimator
+from sapsan.lib.estimator.torch_backend import TorchBackend
 from sapsan.lib.data import get_loader_shape
 
 
@@ -42,8 +42,8 @@ class PICAEModel(torch.nn.Module):
         self.total_layers = self.encoder_nlayers + self.decoder_nlayers
         self.outlayer_padding = kernel_size[0] // 2, kernel_size[1] // 2, kernel_size[2] // 2
         
-        self.te = TorchEstimator
-        self.device = self.te.set_device(self) 
+        self.tb = TorchBackend
+        self.device = self.tb.set_device(self) 
 
 
         ############## Define Encoder layers
@@ -58,7 +58,7 @@ class PICAEModel(torch.nn.Module):
                 cell_outputsize = self.nfilters
                 stridelen = 2
        
-            encoder_cell_list.append(self.te.to_device(self, torch.nn.Conv3d(
+            encoder_cell_list.append(self.tb.to_device(self, torch.nn.Conv3d(
                                                              out_channels = cell_outputsize, 
                                                              in_channels = cell_inpsize, 
                                                              kernel_size = self.kernel_size, 
@@ -81,7 +81,7 @@ class PICAEModel(torch.nn.Module):
                 cell_padding = 0
                 stridelen = 2        
             
-            decoder_cell_list.append(self.te.to_device(self, torch.nn.ConvTranspose3d(
+            decoder_cell_list.append(self.tb.to_device(self, torch.nn.ConvTranspose3d(
                                                               out_channels = cell_outputsize, 
                                                               in_channels = cell_inpsize,
                                                               kernel_size = self.kernel_size, 
@@ -106,7 +106,7 @@ class PICAEModel(torch.nn.Module):
         self.ddzKernel[1,1,2] = 0.5
         #### declare weights
         self.weights = torch.zeros((self.output_size, self.input_size, 3, 3, 3))
-        self.weights = self.weights.type(self.te.tensor_to_device(self))
+        self.weights = self.weights.type(self.tb.tensor_to_device(self))
         #dfy/dx
         self.weights[0,0,::] = torch.zeros(3, 3, 3)
         self.weights[0,1,::] = self.ddxKernel.clone()
@@ -137,16 +137,16 @@ class PICAEModel(torch.nn.Module):
         with torch.no_grad():
             self.curlConv.weight = torch.nn.Parameter(self.weights)
         self.curlField = torch.zeros([self.batch,self.input_size,self.il,self.jl,self.kl])
-        self.curlField = self.curlField.type(self.te.tensor_to_device(self))
+        self.curlField = self.curlField.type(self.tb.tensor_to_device(self))
         self.register_buffer('r_curlField', self.curlField)
         self.curlGrad = torch.zeros([self.batch,self.output_size,self.il,self.jl,self.kl])
-        self.curlGrad = self.curlGrad.type(self.te.tensor_to_device(self))    
+        self.curlGrad = self.curlGrad.type(self.tb.tensor_to_device(self))    
         self.register_buffer('r_curlGrad', self.curlGrad)        
         
     def padHITperiodic(self,field):
         oldSize = field.size(-1)
         newSize = oldSize + 3 #2nd order accurate periodic padding at boundary
-        newField = torch.zeros(self.batch,self.input_size,newSize,newSize,newSize).type(self.te.tensor_to_device(self))
[…]
