### 2021-08-21 4f8aa0a
Added model.forward() to be recorded by MLflow
	- if torch is used
