### 2021-08-24 d5e919b
Loaded models can continue to be trained
	- reworked how the models are initialized
	  this had large implication for the overall workflow
	- model initializaiton requires to provide "loaders" now
	- optimizer dict state is correctly saved and loaded
	- optimizer state is moved to cpu or gpu depending on a setup
	  (catalyst doesn't do it on its own, causing issues when
	   evaluating a loaded model)
	- Loaders for Train and Evaluate now have the same format
	  and in general the functions have an identical interface
	  for both pytorch and sklearn
	- Since the loaded models can now continue t
[…]
