+++ sapsan/lib/data/data_functions.py
@@ -8,7 +8,7 @@ from sapsan.core.models import Dataset, DatasetPlugin
      
 
 def torch_splitter(x, y, 
-                   batch_num: int,
+                   batch_num: int = 1,
                    train_fraction = None,
                    shuffle: bool = False):
     
@@ -20,7 +20,6 @@ def torch_splitter(x, y,
         x_train = x_valid = x
         y_train = y_valid = y
 
-    print(np.shape(x_train), np.shape(y_train))
     dataset=TensorDataset(from_numpy(x_train).float(),
                                                     from_numpy(y_train).float())
     train_loader = DataLoader(dataset=TensorDataset(from_numpy(x_train).float(),
@@ -31,9 +30,9 @@ def torch_splitter(x, y,
 
     valid_loader = DataLoader(dataset=TensorDataset(from_numpy(x_valid).float(),
                                                   from_numpy(y_valid).float()),
-                            batch_size=batch_num,
-                            shuffle=shuffle,
-                            num_workers=4)
+                              batch_size=batch_num,
+                              shuffle=shuffle,
+                              num_workers=4)
     print('Train data shapes: ', x_train.shape, y_train.shape)
     print('Valid data shapes: ', x_valid.shape, y_valid.shape)
     
