+++ sapsan/tests/examples/test_examples.py
@@ -9,7 +9,7 @@ class TestExamples(unittest.TestCase):
     """Tests example notebooks."""
     
     def notebooks(self):
-        return ['cnn_example.ipynb', 'picae_example.ipynb','krr_example.ipynb']
+        return ['cnn_example.ipynb', 'picae_example.ipynb', 'krr_example.ipynb']
 
     def test_examples(self):
         """Tests examples"""       
