### 2021-08-25 8b06d25
Fixed CLI click initialization + cleaned up estimator names
