+++ sapsan/__init__.py
@@ -1,7 +1,7 @@
 from ._version import __version__
-from .lib.estimator.krr.krr import KRR, KRRConfig
-from .lib.estimator.cnn.spacial_3d_encoder import CNN3d, CNN3dConfig
-from .lib.estimator.picae.picae_encoder import PICAE, PICAEConfig
+from .lib.estimator.krr.krr_estimator import KRR, KRRConfig
+from .lib.estimator.cnn.cnn3d_estimator import CNN3d, CNN3dConfig
+from .lib.estimator.picae.picae_estimator import PICAE, PICAEConfig
 from .lib.estimator.torch_backend import load_estimator, TorchBackend
 from .lib.estimator.sklearn_backend import load_sklearn_estimator, SklearnBackend
 from .lib.experiments.evaluate import Evaluate
