+++ sapsan/lib/estimator/__init__.py
@@ -1,5 +1,5 @@
-from .krr.krr import KRR, KRRConfig
-from .cnn.spacial_3d_encoder import CNN3d, CNN3dConfig
-from .picae.picae_encoder import PICAE, PICAEConfig
+from .krr.krr_estimator import KRR, KRRConfig
+from .cnn.cnn3d_estimator import CNN3d, CNN3dConfig
+from .picae.picae_estimator import PICAE, PICAEConfig
 from .torch_backend import load_estimator, TorchBackend
 from .sklearn_backend import load_sklearn_estimator, SklearnBackend
