### 2021-08-25 98c71b7
"sapsan create --ddp" option copies torch_backend
	- gives ability to cutomize Catalyst Runner
	- adjust DDP settings based on the linked
	  Catalyst DDP tutorial in the Wiki
	- will be useful when running on HPC
