### 2021-08-25 dcb9567
Minor spelling fix in the template
