### 2021-08-26 e67a30d
Save, Load, and continue training (#121)
* Updated eval - runtime MLflow metric logging

* MLflow: evaluate runs will be nested under the recent train run

* Fixed 'nested' key in fake backend

* Fixed save/load models, added sklearn_backend, torch_backend

	- sklearn estimator has a structure close to torch
	- pytorch_estimator -> torch_backend
	- cleared up variable name conventions throughout

* Added model.forward() to be recorded by MLflow

	- if torch is used

* Loaded models can continue to be trained

	- reworked how the models are initialized
	  this had large implication for the over
[…]
