+++ sapsan/examples/GUI/st_experiments.py
@@ -1,38 +1,36 @@
-import streamlit as st
 import os
 import sys
 import inspect
+import time
+import json
+import signal
+import numpy as np
 from pathlib import Path
+from collections import OrderedDict
+
+import pandas as pd
+import matplotlib as mpl
+import matplotlib.pyplot as plt
+import plotly.express as px
+import configparser
+import webbrowser
+from io import BytesIO
+from threading import Thread
+#from st_state_patch import SessionState
+#from multiprocessing import Process
+
+import torch
+import streamlit as st
+from streamlit.report_thread import add_report_ctx
 
 #uncomment if cloned from github!
 sys.path.append(str(Path.home())+"/Sapsan/")
 
-from sapsan.lib.backends.fake import FakeBackend
-from sapsan.lib.backends.mlflow import MLflowBackend
+from sapsan.lib.backends import FakeBackend, MLflowBackend
 from sapsan.lib.data import HDF5Dataset, EquidistantSampling, flatten
-from sapsan.lib.estimator import CNN3d, CNN3dConfig
-from sapsan.lib.estimator.cnn.spacial_3d_encoder import CNN3dModel
-from sapsan.lib.experiments.evaluate import Evaluate
-from sapsan.lib.experiments.train import Train
-from sapsan.utils.plot import model_graph
-
-import pandas as pd
-import torch
-import matplotlib.pyplot as plt
-import configparser
-import webbrowser
-import time
-import numpy as np
-from threading import Thread
-from streamlit.ReportThread import add_report_ctx
-import json
-from collections import OrderedDict
-import plotly.express as px
-import os
-import signal
-import sys
-from st_state_patch import SessionState
-from multiprocessing import Process
+from sapsan import Train, Evaluate, CNN3d, CNN3dConfig, model_graph
+from sapsan.lib.estimator.cnn.cnn3d_estimator import CNN3dModel
+from sapsan.utils.plot import model_graph, pdf_plot, cdf_plot, slice_plot, plot_params
 
 #initialization of defaults
 cf = configparser.RawConfigParser()
@@ -77,10 +75,10 @@ Sapsan has a BSD-style license, as found in the [LICENSE] (https://github.com/pi
         """
         )
         
-            
+    
 def cnn3d():
     st.title('Sapsan Configuration')
-    st.write('This demo is meant to present capabilities of Sapsan. You can configure each part of the experiment at the sidebar. Once you are done, you can see the summary of your runtime parameters under _Show configuration_. In addition you can review the model that is being used (in the custom setup, you will also be able to edit it). Lastly click the _Run experiment_ button to train the test the ML model.')
+    st.write('This demo is meant to present capabilities of Sapsan. You can configure each part of the experiment at the sidebar. Once you are done, you can see the summary of your runtime parameters under _Show configuration_. In addition you can review the model that is being used (in the custom setup, you will also be able to edit it). Lastly click the _Run experiment_ button to train the test the ML model.')    
     
     st.sidebar.markdown("General Configuration")
     
@@ -205,7 +203,12 @@ def cnn3d():
                         
             
             time.sleep(0.1) 
-            
+
+    def plot_static():
+        buf = BytesIO()
+        plt.savefig(buf, format="png",  dpi=50)
+        st.image(buf)
+    
     def load_data(checkpoints):
         #Load the data      
         features = widget_values['features'].split(',')
@@ -216,15 +219,15 @@ def cnn3d():
         
         checkpoints = np.array([int(i) for i in checkpoints.split(',')])
         
+        #print('--ST Shapes--', batch_size, input_size)
         data_loader = HDF5Dataset(path=widget_values['path'],
-                           features=features,
-                           target=target,
-                           checkpoints=checkpoints,
-                           batch_size=text_to_list(widget_values['batch_size']),
[…]
