+++ requirements.txt
@@ -2,9 +2,9 @@ numpy>=v1.19.2
 Click>=7.1.2
 catalyst>=21.5
 h5py>=2.10.0
-notebook>6.4.3
+notebook>=6.4.3
 jupyter>=1.0.0
-tornado>6.1.0
+tornado>=6.1.0
 matplotlib>=3.3.2
 mlflow>=1.18.0
 pandas>=1.1.0
