### 2021-08-26 3935098
Plotting, GUI, Data Loaders, Requirements  (#122)
* Plotting, GUI, Data Loaders, Requirements

	- Plotting
		- plotting routines return Axes object
		- all pars are changed for the axes instead of plt
		  which allows individual tweaking
		- figsize and ax arguments added to most plotting routines
		- PDF and CDF are plotted as one figure for eval
		- Universal plotting params expanded and
		  were made easily accessible through plot_params()

	- GUI
		- up to date with Streamlit 0.87.0 and Sapsan 0.3.1
		- PDF and CDF plots are now showed as well
		- fixed data loading issue in regards to tra
[…]
