+++ requirements.txt
@@ -2,9 +2,9 @@ numpy>=v1.19.2
 Click>=7.1.2
 catalyst>=21.5
 h5py>=2.10.0
-notebook<6.1.6
+notebook>=6.4.3
 jupyter>=1.0.0
-tornado<6.0.0
+tornado>=6.1.0
 matplotlib>=3.3.2
 mlflow>=1.18.0
 pandas>=1.1.0
@@ -17,7 +17,7 @@ scipy>=1.5.2
 seaborn>=0.11.1
 six>=1.15.0
 graphviz>=0.14
-streamlit==0.58.0
+streamlit>=0.87.0
 opencv-python>=4.5.1
 jupytext>=1.11
 pytest>=6.2

[…]
