### 2021-08-27 a1b9e5a
PDF and CDF plot are recorded as a double figure in MLflow
