### 2021-08-27 cf26698
Fixed: n_output_channel in Eval, Sampling Warning, Tests
