### 2021-08-27 59d8f63
Data loader and Eval accept data without target + Minor (#123)
* Data loader and Eval accept data without target

	- useful when there is no ground truth to compare to
	- will still output pdf, cdf, and spatial,
	  without comparison metrics of course
	- eval now output a dict of "pred_cube",
	  with "target_cube" if such is provided

* PDF and CDF plot are recorded as a double figure in MLflow

* Edited the examples to reflect the Eval output changes

* Fixed: n_output_channel in Eval, Sampling Warning, Tests

Co-authored-by: Platon Karpov <plkarpov@ucsc.edu>
Co-authored-by: pikarpov-LANL <pi
[…]
