+++ sapsan/lib/experiments/evaluate.py
@@ -4,7 +4,7 @@ evaluation_experiment = Evaluate(backend=tracking_backend,
                                  model=training_experiment.model,
                                  data_parameters = data_loader)
 
-target_cube, pred_cube = evaluation_experiment.run()
+cubes = evaluation_experiment.run()
 """
 
 import os
@@ -22,7 +22,7 @@ from sapsan.utils.plot import pdf_plot, cdf_plot, slice_plot, plot_params
 from sapsan.utils.shapes import combine_cubes, slice_of_cube
 
 class Evaluate(Experiment):
-    def __init__(self,                 
+    def __init__(self,
                  model: Estimator,
                  data_parameters,
                  backend = FakeBackend(),
@@ -37,20 +37,26 @@ class Evaluate(Experiment):
         self.batch_num = self.data_parameters.batch_num
         self.cmap = cmap
         self.axis = len(self.input_size)
+        self.targets_given = True
+        self.flat = flat
+        self.artifacts = []        
         
         if type(self.model.loaders) in [list, np.array]:
             self.inputs = self.model.loaders[0]
-            self.targets = self.model.loaders[1]
+            try: 
+                self.targets = self.model.loaders[1]
+            except: 
+                self.targets_given = False
+                print('Warning: no target given; only predicting...')
         else:
-            self.inputs, self.targets = iter(self.model.loaders['train']).next()
-            self.targets = self.targets.numpy()
+            try:
+                self.inputs, self.targets = iter(self.model.loaders['train']).next()
+                self.targets = self.targets.numpy()
+            except: 
+                self.inputs = iter(self.model.loaders['train']).next()[0]
+                self.targets_given = False
+                print('Warning: no target given; only predicting...')
         
-        self.flat = flat
-        if self.flat: self.n_output_channels = self.targets.shape[0] #flat arrays don't have batches
-        elif len(self.targets.shape)<(self.axis+2): self.n_output_channels = 1
-        else: self.n_output_channels = self.targets.shape[1]
-
-        self.artifacts = []
         
     def get_metrics(self) -> Dict[str, float]:
         if 'train' in self.model.metrics():
@@ -81,35 +87,36 @@ class Evaluate(Experiment):
         start = time.time()
         
         self.backend.start('evaluate', nested = True)
-
-        pred = self.model.predict(self.inputs, self.model.config)
         
+        pred = self.model.predict(self.inputs, self.model.config)              
+
         end = time.time()
         runtime = end - start
         self.backend.log_metric("eval - runtime", runtime)
+
+        #determine n_output_channels form prediction
+        if self.flat:
+            self.n_output_channels = int(np.around(
+                                     np.prod(pred.shape)/np.prod(self.inputs.shape[1:])
+                                     )) #flat arrays don't have batches
+        elif len(pred.shape)<(self.axis+2): 
+            self.n_output_channels = int(np.around(
+                                         np.prod(pred.shape[1:])/np.prod(self.inputs.shape[2:])
+                                         ))
+        else: self.n_output_channels = pred.shape[1]            
         
-        mpl.rcParams.update(plot_params())
+        if self.targets_given: 
+            series = [pred, self.targets]
+            names = ['predict', 'target']
+        else: 
+            series = [pred]
+            names = ['predict']
         
-        fig = plt.figure(figsize=(12,6), dpi=60)
-        (ax1, ax2) = fig.subplots(1,2)
-
-        pdf = pdf_plot([pred, self.targets], names=['predict', 'target'], ax=ax1)
-        plt.savefig("pdf_plot.png")
-
-        self.artifacts.append("pdf_plot.png")
-
[…]
