### 2021-08-28 pikarpov-LANL/Sapsan#99 @kburns
Thanks! I downloaded the latest (v0.4.0) and tried to run the tests, but it failed saying

```
FileNotFoundError: [Errno 2] No such file or directory: '/Users/kburns/Software/miniconda3/envs/sapsan38/lib/python3.8/site-packages/sapsan/examples/cnn_example.ipynb'
```

I think the issue is that `setuptools.find_packages()` only looks for / includes `*.py` files, so the entire examples directory and all the `*.ipynb` don't get included in the builds for PyPI.
