### 2021-08-28 3921d46
Fixed 'get_examples' command
