### 2021-08-28 pikarpov-LANL/Sapsan#99 @pikarpov-LANL
Thank you! I forgot to test the PyPi release locally, better write a test for it. I just fixed the issue in `v0.4.1`, please check it out.
