+++ MANIFEST.in
@@ -1,1 +1,1 @@
-include examples/*
+include sapsan/examples/*
