+++ MANIFEST.in
@@ -1,1 +1,2 @@
 include sapsan/examples/*
+include sapsan/examples/data/t0/*
