+++ sapsan/_version.py
@@ -1,1 +1,1 @@
-__version__ = 'v0.4.0'
+__version__ = 'v0.4.3'
