### 2021-08-28 pikarpov-LANL/Sapsan#99 @kburns
Thanks! Tests are passing for me now. Just a few very minor points on the new draft (line numbers refer to the latest whedon-generated PDF):

- [ ] Typo line 86: "Loos" -> "Loss"
- [ ] Typo line 131: "10243" -> "1024^3" (maybe this is just a rendering problem?)
- [ ] For both examples, can you say qualitatively how much filtering was done to produce the training data from the DNS? That's really useful for understanding how much of the dynamics the ML closure is responsible for.
