+++ MANIFEST.in
@@ -1,1 +1,3 @@
-include examples/*
+include sapsan/examples/*
+include sapsan/examples/data/t0/*
+include sapsan/examples/GUI/*
