### 2021-08-30 pikarpov-LANL/Sapsan#99 @pikarpov-LANL
1. Fixed
2. It is a rendering issue - I am not sure what to do since it shows up fine on the paper.md on github (using html <sup>3</sup>, but ^ doesn't work either)
3. Added the exact filtering amount I used
