### 2021-08-30 1a3249d
Added 'run_name' parameter to train and evaluate classes
	- allows user to change the 'run_name' tag
	  as recorded by MLflow
