### 2021-08-30 #3199 @kburns
@dfm my checklist is complete. Thanks to the authors for their patience! I think everything looks great except some superscripts don't seem to be rendering properly in the PDF, (e.g. line 131 in the latest proof).
