### 2021-08-31 pikarpov-LANL/Sapsan#99 @pikarpov-LANL
Glad to hear! Your feedback has certainly been appreciated - the code and the paper are in a much better shape now. I will check with the editor on the superscript rendering issue.
