### 2021-08-31 119d381
Added kwags par to Evaluate that control axes settings for pdf and cdf
	- Example: Evaluate('pdf_xlim'=(0,10))
	  that will set xlim for the pdf plot to (0,10).
	  Internally, ax.set_xlim=(0,10) for the pdf plot
	- Any parameter in the form ax.set_{par} can be passed
	  through Evaluate to adjust pdf and cdf plots
	- Useful when you are doing a large model parameter search
	  and want pdf and cdf to have similar bounds for easier
	  visual comparison
