+++ sapsan/lib/experiments/evaluate.py
@@ -28,7 +28,8 @@ class Evaluate(Experiment):
                  backend = FakeBackend(),
                  cmap: str = 'plasma',
                  flat: bool = False,
-                 run_name: str = 'evaluate'):
+                 run_name: str = 'evaluate',
+                 **kwargs):
         self.model = model
         self.backend = backend
         self.experiment_metrics = dict()
@@ -41,7 +42,11 @@ class Evaluate(Experiment):
         self.targets_given = True
         self.flat = flat
         self.run_name = run_name
-        self.artifacts = []        
+        self.artifacts = []    
+        self.axes_pars = ['pdf_xlim','pdf_ylim',
+                          'cdf_xlim','cdf_ylim']
+        self.kwargs = kwargs
+        
         
         if type(self.model.loaders) in [list, np.array]:
             self.inputs = self.model.loaders[0]
@@ -192,7 +197,12 @@ class Evaluate(Experiment):
         (ax1, ax2) = fig.subplots(1,2)
 
         pdf = pdf_plot(series, names=names, ax=ax1)
+        self.set_axes_pars(pdf)
+        
         cdf = cdf_plot(series, names=names, ax=ax2)
+        cdf = self.set_axes_pars(cdf)
+        
+        plt.tight_layout()
         plt.savefig("pdf_cdf.png")
         self.artifacts.append("pdf_cdf.png")                        
 
@@ -208,8 +218,14 @@ class Evaluate(Experiment):
                 slice_names.append(key)
                       
         slices = slice_plot(slice_series, names=slice_names, cmap=self.cmap)
+        plt.tight_layout()
         plt.savefig("slices_plot.png")
         self.artifacts.append("slices_plot.png")
                       
         return slices_cubes
-        
+    
+    def set_axes_pars(self, ax):
+        for key in self.axes_pars:
+            if key in self.kwargs.keys(): 
+                axes_attr = getattr(ax, 'set_'+key.split('_')[-1])
+                axes_attr(self.kwargs[key])                        
