### 2021-09-03 eea7ccf
Added interface to record extra metrics, etc. in MLflow
	- they can be added straight in the Jupyter notebook
	  after evaluation step (or train)
	- both Evaluation and Train have a 'run_id' parameter
	- 'run_id' will need to be passed to backend.resume()
	  to restart the run
	- will update MLflow instructions on Wiki
