### 2021-09-03 af30799
Added `ks` parameter to cdf_plot()
	- if ks=True, kolmogorov-smirnov statisic will be computed
	  and returned as a second argument along with ax

	  ax, ks = cdf_plot()
