### 2021-09-05 pikarpov-LANL/Sapsan#100 @MilesCranmer
Hi @pikarpov,

I sincerely apologize for the late followup. I have been on an internship where all the code writing is internal, and have been bad at going through my GitHub notifications, and this dropped off my radar. I am very sorry about that, and thanks for your patience.

The updated repo, documentation, and paper look significantly better. Fantastic work! This is a really great project that I am quite excited that you and your coauthors have created. I am very happy with how this work has turned out. The docs are really really nice, and I can clearly see how to implement different models. Thanks for looking into getting the colab working too - it's a time-consuming process to get that working, but it will be really useful for others in quickly trying out your package, so I think it will be worth it. I confirmed that it works as intended!

I think this will be a very useful repo for the community :) The README is very clear about the goals of the package, and I think they well-represent the framework. I think that updated focus will be very helpful in attracting others to work in this intersection of fields.

Cheers!
Miles
