### 2021-09-05 #3199 @MilesCranmer
Hi @dfm @pikarpov-LANL and co-authors,

I sincerely apologize for the late followup. I have been on an internship where all the code writing is internal, and have been bad at going through my GitHub notifications, and this dropped off my radar. I am very sorry about that, and thanks for your patience! (and thanks @dfm for the helpful ping :) )

All of my comments have been addressed - see also my response in the issue - and all checkpoints above are hit. I am very happy with the Sapsan project! It's quite a nice framework, and the plethora of documentation available is fantastic - perfect for the intended audience(s).

Cheers,
Miles
