### 2021-09-05 4d21954
Add wheel install to unit test
