+++ .github/workflows/CI.yml
@@ -37,6 +37,7 @@ jobs:
       - name: "Install Sapsan"
         run: |
             python -m pip install --upgrade pip
+            pip install wheel
             pip install -r requirements.txt
             python setup.py install
       - name: "Tests"
