### 2021-09-09 pikarpov-LANL/Sapsan#100 @pikarpov-LANL
Hi Miles,

Thank you for the kind words! A lot of work has gone into `Sapsan`. My collaborators and I appreciated both yours and Keaton's comments that helped improve the framework. We always welcome constructive criticism and suggestions!
