### 2021-09-09 03fc26d
Fixed mlflow.close_active_run() to run until there are no runs
	- added while loop for mlflow.end_run()
