### 2021-09-09 f006b5f
MLflow: adding new parameters after the run (#130)
* Added GUI to pypi distro

* Added 'run_name' parameter to train and evaluate classes

	- allows user to change the 'run_name' tag
	  as recorded by MLflow

* Fixed moving tensors to specific cuda id during evaluation

* Added kwags par to Evaluate that control axes settings for pdf and cdf

	- Example: Evaluate('pdf_xlim'=(0,10))
	  that will set xlim for the pdf plot to (0,10).
	  Internally, ax.set_xlim=(0,10) for the pdf plot
	- Any parameter in the form ax.set_{par} can be passed
	  through Evaluate to adjust pdf and cdf plots
	- Useful 
[…]
