+++ requirements.txt
@@ -6,7 +6,7 @@ notebook>=6.4.3
 jupyter>=1.0.0
 tornado>=6.1.0
 matplotlib>=3.3.2
-mlflow>=1.18.0
+mlflow>=1.20.1
 pandas>=1.1.0
 Pillow>=8.1.0
 plotly>=5.2.0
