### 2021-09-09 85343b6
Removed unnecessary GUI files
