+++ sapsan/examples/GUI/st_custom.py
@@ -1,425 +1,0 @@
-import streamlit as st
-import os
-import sys
-import inspect
-from pathlib import Path
-
-#uncomment if cloned from github!
-sys.path.append(str(Path.home())+"/Sapsan/")
-
-from sapsan.lib.backends import FakeBackend, MLflowBackend
-from sapsan.lib.data import HDF5Dataset, EquidistantSampling, flatten
-from sapsan import Train, Evaluate, CNN3d, CNN3dConfig, model_graph
-from sapsan.lib.estimator.cnn.cnn3d_estimator import CNN3dModel
-from sapsan.utils.plot import model_graph
-
-import pandas as pd
-import torch
-import matplotlib.pyplot as plt
-import configparser
-import webbrowser
-import time
-import numpy as np
-from threading import Thread
-from streamlit.report_thread import add_report_ctx
-import json
-from collections import OrderedDict
-import plotly.express as px
-import os
-import signal
-import sys
-from st_state_patch import SessionState
-from multiprocessing import Process
-
-cf = configparser.RawConfigParser()
-widget_values = {}
-
-def custom():
-    st.title('Sapsan Configuration')
-    st.write('This demo is meant to present capabilities of Sapsan. You can configure each part of the experiment at the sidebar. Once you are done, you can see the summary of your runtime parameters under _Show configuration_. In addition you can review the model that is being used (in the custom setup, you will also be able to edit it). Lastly click the _Run experiment_ button to train the test the ML model.')
-    
-    st.sidebar.markdown("General Configuration")
-    
-    try:
-        cf.read('temp.txt')
-        temp = dict(cf.items('config'))
-    except: pass
-    
-    def make_recording_widget(f):
-        """Return a function that wraps a streamlit widget and records the
-        widget's values to a global dictionary.
-        """
-        def wrapper(label, *args, **kwargs):
-            widget_value = f(label, *args, **kwargs)
-            widget_values[label] = widget_value
-            return widget_value
-
-        return wrapper
-           
-    def widget_history_checkbox(title, params):
-        if st.sidebar.checkbox(title):
-            widget_history_checked(params)
-        else:
-            widget_history_unchecked(params)
-    
-    def widget_history_checked(params):
-        widget_type = {number:int, number_main:int, text:str, text_main:str, checkbox:bool}
-        for i in range(len(params)):
-            label = params[i]['label']
-            default = params[i]['default']
-            widget = params[i]['widget']
-
-            not_widget_params = ['default', 'widget', 'widget_type']
-            additional_params = {key:value for key, value in params[i].items() if key not in not_widget_params}
-            try:
-                if widget_values[label+'_flag'] == True:
-                    widget_values[label+'_flag'] = False
-                    try:
-                        widget_values[label+'_default'] = widget_type[widget](temp[label])
-                        widget(value = widget_type[widget](temp[label]), **additional_params)
-                    except: widget(value = widget_values[label+'_default'], **additional_params)
-                else:
-                    widget(value = widget_values[label+'_default'], **additional_params)
-            except: 
-                widget_values[label+'_flag'] = False
-                widget(value = widget_type[widget](default), **additional_params)
-    
-    def widget_history_unchecked(params):
-        widget_type = {number:int, number_main:int, text:str, text_main:str, checkbox:bool}
-        for i in range(len(params)):
-            label = params[i]['label']
-            default = params[i]['default']
-            widget = params[i]['widget']
-            
-            widget_values[label+'_flag'] = True        
-            widget_values[label+'_default'] = widget_type[widget](default)
-
-    def load_config(config_file):
-        cf.read(config_file)
-        config = dict(cf.items('sapsan_config'))
-        return config
-    
[…]
