### 2021-09-10 14c1a11
CLI, GUI, Model Graph
	- CLI: GUI will be copied with examples
	- GUI: simplified file structure, fixed up compatibility issues
	- Model Graph: removed automatic re-setting the number of
channels to 1
