### 2021-09-10 62e429f
CLI, GUI, Model Graph (#132)
* Removed unnecessary GUI files

* Locked requirement streamlit == 0.84.2

	- there is a known bug causing DataFrames to fail
	- will unlock once resolved

* CLI, GUI, Model Graph

	- CLI: GUI will be copied with examples
	- GUI: simplified file structure, fixed up compatibility issues
	- Model Graph: removed automatic re-setting the number of channels to 1
