### 2021-09-10 7ac9409
welcome page for GUI example (#133)
* Added a welcome page for GUI
