+++ docs/joss/paper.md
@@ -20,7 +20,7 @@ authors:
     orcid: 0000-0003-2624-0056
     affiliation: 2
 affiliations:
-  - name: Department of Astornomy & Astrophysics, University of California, Santa Cruz, CA
+  - name: Department of Astronomy & Astrophysics, University of California, Santa Cruz, CA
     index: 1
   - name: Los Alamos National Laboratory, Los Alamos, NM
     index: 2
@@ -31,105 +31,92 @@ bibliography: paper.bib
 ---
 
 # Summary
-[Sapsan](https://github.com/pikarpov-LANL/Sapsan) is a framework to make Machine Learning (ML) more accessible in the study of turbulence, focusing on astrophysical applications. Sapsan includes modules to load, filter, subsample, batch, and split the data from hydrodynamic (HD) simulations for training and validation. Next, the framework includes built-in conventional and physics-informed published estimators that were used for turbulence modeling. This ties into Sapsan's custom estimator module, aimed at designing a custom ML model layer-by-layer, which is the core benefit of using the framework. To share your custom model, every new project created via Sapsan comes with pre-filled, ready-for-release Docker files. Furthermore, training and evaluation modules come with Sapsan as well. The latter, among other features, includes the construction of power spectra and comparison to established analytical turbulence closure models, such as a gradient model. Thus, Sapsan takes out all the hard work from data preparation and analysis, leaving one focused on the ML model design itself.
+[``Sapsan``](https://github.com/pikarpov-LANL/Sapsan) is a framework designed to make Machine Learning (ML) more accessible in the study of turbulence, with a focus on astrophysical applications. ``Sapsan`` includes modules to load, filter, subsample, batch, and split the data from hydrodynamic (HD) simulations for training and validation. Next, the framework includes built-in conventional and physically-motivated estimators that have been used for turbulence modeling. This ties into ``Sapsan``'s custom estimator module, aimed at designing a custom ML model layer-by-layer, which is the core benefit of using the framework. To share your custom model, every new project created via ``Sapsan`` comes with pre-filled, ready-for-release Docker files. Furthermore, training and evaluation modules come with ``Sapsan`` as well. The latter, among other features, includes the construction of power spectra and comparison to established analytical turbulence closure models, such as a gradient model. Thus, ``Sapsan`` attempts to minimize the hard work required for data preparation and analysis, leaving one to focus on the ML model design itself.
 
 # Statement of Need
 
-It has been challenging for domain sciences to adopt Machine Learning (ML) for their respective projects, particularly for physical simulations modeling turbulence. It is rather challenging to prove that an ML model has arbitrarily learned the laws of physics embedded into the problem with the ability to extrapolate within the parameter-space of the simulation. The inability to directly infer the prediction capabilities of ML is one of the major causes behind the slow adaption rates; however, the community cannot ignore the effectiveness of ML.
+Domain sciences have been slow to adopt Machine Learning (ML) for a range of projects, but particularly for physical simulations modeling turbulence. It is challenging to prove that an ML model has learned the laws of physics in a particular problem, and that it has the ability to extrapolate within the parameter-space of the simulation. The inability to directly infer the predictive capabilities of ML is one of the major causes behind the slow adoption rates; however, the community cannot ignore the effectiveness of ML.
 
[…]
