+++ docs/joss/paper.bib
@@ -156,7 +156,7 @@ archivePrefix = {arXiv},
        author = {{Carleo}, Giuseppe and {Cirac}, Ignacio and {Cranmer}, Kyle and
          {Daudet}, Laurent and {Schuld}, Maria and {Tishby}, Naftali and
          {Vogt-Maranto}, Leslie and {Zdeborov{\'a}}, Lenka},
-        title = "{Machine learning and the physical sciences*}",
+        title = "{Machine learning and the physical sciences}",
       journal = {Reviews of Modern Physics},
      keywords = {Physics - Computational Physics, Astrophysics - Cosmology and Nongalactic Astrophysics, Condensed Matter - Disordered Systems and Neural Networks, High Energy Physics - Theory, Quantum Physics},
          year = 2019,
