### 2021-09-17 #3199 @pikarpov-LANL
@dfm thank you for the edits! I reviewed and accepted the pull request. I would like to update the the figures a little bit to be closer to the plots from my Astrophysical Journal paper; text will remain the same. Let me just run the final version by my collaborators, and then I will go through the last steps you outlined.
