### 2021-09-17 #3199 @dfm
@pikarpov-LANL: great! So, is there a parallel ApJ paper to this one? What is the current status of that review?
