### 2021-09-17 #3199 @pikarpov-LANL
@dfm Yes, there is. We are about to submit it for review; I will include the appropriate reference. It barely includes any details on `Sapsan` itself, besides mentioning it. Results/plot wise the paper is much more extensive; JOSS paper just includes a few examples.
