### 2021-09-22 66dfcd2
Added all plot and filters to __init__
