+++ sapsan/utils/__init__.py
@@ -1,1 +1,3 @@
 from .physics import PowerSpectrum, GradientModel, DynamicSmagorinskyModel, picae_func
+from .plot import plot_params, pdf_plot, cdf_plot, slice_plot, line_plot, model_graph, log_plot
+from .filters import spectral, box, gaussian
