### 2021-09-27 3988137
Fixed up Gradient Model calculation consistency
