### 2021-10-05 f745ceb
Returned auto-termination of MLflow tracking after Evaluation.run()
