### 2021-10-12 f738089
Fixed derivative multiplication for Gradient Model
