### 2021-10-12 6df778b
Updated spectrum_plot formatting, renamed, default grid for plots
	- Renamed: plot_spectrum() -> spectrum_plot()
	- 	   tensor() -> ReynoldsStress()
