### 2021-10-12 f43c37f
Gradient model fix, plot formatting, MLflow auto-end evaluation runs (#136)
* Added all plot and filters to __init__

* Fixed up Gradient Model calculation consistency

* Returned auto-termination of MLflow tracking after Evaluation.run()

* Fixed derivative multiplication for Gradient Model

* Updated spectrum_plot formatting, renamed, default grid for plots

	- Renamed: plot_spectrum() -> spectrum_plot()
	- 	            tensor() -> ReynoldsStress()
