### 2021-10-20 eba8418
Changed --ddp to --gtb or --get_torch_backend
	- Added new command: sapsan get_torch_backend
	- Editting the runner is not only needed for DDP,
	  but lots of applications
