### 2021-10-20 b4c6ec4
Fixed DDP + CLI "--get_torch_backend" (#137)
* Fixed DistributedDataParallel (DDP) + run info printing

	- engine is no longer overwritten;
	  will be determined automatically by Catalist

* Changed --ddp to --gtb or --get_torch_backend

	- Added new command: sapsan get_torch_backend
	- Editting the runner is not only needed for DDP,
	  but lots of applications
