+++ README.md
@@ -1,17 +1,17 @@
 # Sapsan  <a href="http://sapsan.app"><img src="https://github.com/pikarpov-LANL/Sapsan/blob/images/docs/images/logo3_black_slim_notitle_whitebg.png?raw=true"  alt="Sapsan logo" align="right" width="100"></a>
 
-Sapsan is a pipeline for Machine Learning (ML) based turbulence modeling. While turbulence is important in a wide range of mediums, the pipeline primarily focuses on astrophysical application. With Sapsan, one can create their own custom models or use either conventional or physics-informed ML approaches for turbulence modeling included with the pipeline ([estimators](https://github.com/pikarpov-LANL/Sapsan/wiki/Estimators)). Sapsan is designed to take out all the hard work from data preparation and analysis, leaving you focused on ML model design, layer by layer.
+Sapsan is a pipeline for Machine Learning (ML) based turbulence modeling. While turbulence is important in a wide range of mediums, the pipeline primarily focuses on astrophysical applications. With Sapsan, one can create their own custom models or use either conventional or physics-informed ML approaches for turbulence modeling included with the pipeline ([estimators](https://github.com/pikarpov-LANL/Sapsan/wiki/Estimators)). Sapsan is designed to take out all the hard work from data preparation and analysis, leaving you focused on ML model design, layer by layer.
 
-Feel free to check out a website version at [sapsan.app](http://sapsan.app). The interface is indentical to the GUI of the local version of Sapsan, except lacking the ability to edit the model code on the fly.
+Feel free to check out a website version at [sapsan.app](http://sapsan.app). The interface is identical to the GUI of the local version of Sapsan, except lacking the ability to edit the model code on the fly.
 
 ## [Sapsan's Wiki](https://github.com/pikarpov-LANL/Sapsan/wiki)
 
-Please refer to Sapsan's github wiki to learn more about framework's details and capabilities.
+Please refer to Sapsan's Github Wiki to learn more about the framework's details and capabilities.
 
 ## Quick Start
 
 ### 1. Install PyTorch (prerequisite)
-Sapsan can be run on both cpu and gpu. Please follow the instructions on [PyTorch](https://pytorch.org/get-started/locally/) to install the latest version (torch>=1.7.1 & CUDA>=11.0).
+Sapsan can be run on both CPU and GPU. Please follow the instructions on [PyTorch](https://pytorch.org/get-started/locally/) to install the latest version (torch>=1.7.1 & CUDA>=11.0).
 
 ### 2. Install via pip (recommended)
 ```
@@ -43,11 +43,11 @@ sapsan get_examples
 This will create a folder `./sapsan_examples` with appropriate example jupyter notebooks.
 
 ### 5. Create Custom Projects!
-To start a custom project, designing your own custom estimator, i.e. network, go ahead and run:
+To start a custom project, designing your own custom estimator, i.e., network, go ahead and run:
 ```
 sapsan create {name}
 ```
-where `{name}` should be replaced with your custom project name. As a result, a pre-filled template for the estimator, jupyter notebook to run everything from, and Docker will be initialized.
+where `{name}` should be replaced with your custom project name. As a result, a pre-filled template for the estimator, a jupyter notebook to run everything from, and Docker will be initialized.
 
 
 
