### 2021-10-21 #3199 @pikarpov-LANL
@dfm I just pushed the latest version of Sapsan

1. I need to update the plots to reflect the latest changes in analytical subgrid models included with Sapsan + latest formatting. I will get back to you next week, as I am currently on travel.
2. Sapsan v0.4.5
3. DOI: 10.5281/zenodo.5587205
