+++ docs/joss/paper.bib
@@ -272,3 +272,12 @@ url={https://doi.org/10.1038/nature12128}
 	journal = {Science}
 }
 
+@article{liu_meneveau_katz_1994, 
+	 title={On the properties of similarity subgrid-scale models as deduced from measurements in a turbulent jet}, 
+	 volume={275}, 
+	 DOI={10.1017/S0022112094002296}, 
+	 journal={Journal of Fluid Mechanics}, 
+	 publisher={Cambridge University Press}, 
+	 author={Liu, Shewen and Meneveau, Charles and Katz, Joseph}, 
+	 year={1994}, 
+	 pages={83–119}}
