+++ docs/joss/paper.md
@@ -88,33 +88,32 @@ The following is a list of the core functional dependencies[^deps] and a short d
 
 # Applications
 
-While ``Sapsan`` is designed to be highly customizable for a wide variety of projects in the physical sciences, it is optimized for the study of turbulence. 
+While ``Sapsan`` is designed to be highly customizable for a wide variety of projects in the physical sciences, it is optimized for the study of turbulence. In the next section we will demostrate various capabilities of Sapsan working with 2D and 3D data, various machine learning libraries, and built-in analytical tools.
 
 ## Hydro simulations
 
-Here are a few examples of a turbulence closure model trained on the high-resolution Johns Hopkins Turbulence Database [JHTDB, @jhtdb2008]. The dataset used in this comparison is a direct numerical simulation (DNS) of a statistically-stationary isotropic 3D MHD turbulence dataset, $1024^3$ in spatial resolution and covering roughly one large eddy turnover time over 1024 checkpoints, i.e. the dynamical time of the system
-[@Eyink2013]. We compare it with a commonly used Dynamic Smagorinsky (DS) turbulence closure model [@lilly1966]. On the ``Sapsan`` side, a Kernel Ridge Regression model [@murphy2004] is used to demonstrate the effectiveness of conventional ML approaches in tackling turbulence problems. In this test, we used the following setup:
+Here is an examples of a turbulence closure model trained on the high-resolution Johns Hopkins Turbulence Database [JHTDB, @jhtdb2008]. The dataset used in this comparison is a 2D slice of a direct numerical simulation (DNS) of a statistically-stationary isotropic 3D MHD turbulence dataset, $1024^3$ in spatial resolution and covering roughly one large eddy turnover time over 1024 checkpoints, i.e. the dynamical time of the system [@Eyink2013]. We compare it with a commonly used Dynamic Smagorinsky (DS) turbulence closure model [@lilly1966]. On the ``Sapsan`` side, a Kernel Ridge Regression model [@murphy2004] by the means of `scikit-learn` is used to demonstrate the effectiveness of conventional ML approaches in tackling turbulence problems. In this test, we used the following setup:
 
-* __Train features:__ velocity (*u*), vector potential (*A*), magnetic field (*B*), and their respective derivatives at time step = 1. All quantities have been filtered down to 15 Fourier modes to remove small-scale perturbations, mimicking the lower fidelity of a non-DNS simulation.
+* __Train features:__ velocity (*u*), vector potential (*A*), magnetic field (*B*), and their respective derivatives at time step = 1. All quantities have been filtered down to 15 Fourier modes to remove small-scale perturbations, mimicking the lower fidelity of a non-DNS simulation. Next they were sampled down to $128^3$, with the last step leaving a single slice of $128^2$ ready for training.
 * __Model Input:__ low fidelity velocity (*u*), vector potential (*A*), magnetic field (*B*), and their respective derivatives at a set time step in the future.
-* __Model Output:__ velocity stress tensor ($\tau$) at the matching time step in the future, which effectively represents the difference between large and small scale structures of the system.
+* __Model Output:__ velocity stress tensor component ($\tau_{xy}$) at the matching time step in the future, which effectively represents the difference between large and small scale structures of the system.
 
 In Figure 2, it can be seen that the ML-based approach significantly outperforms the DS subgrid model in reproducing the probability density function, i.e., a statistical distribution of the stress tensor. The results are consistent with [@king2016].
 
[…]
