+++ docs/joss/paper.md
@@ -109,11 +109,11 @@ evolving turbulence that is not necessarily isotropic. Turbulence can behave dra
 
 * __Train features:__ velocity (*u*), magnetic field (*B*), and their respective derivatives at time steps before 5 ms (halfway of the total simulation). All quantities have been filtered down with a $\sigma=9$ Gaussian filter to remove small-scale perturbations, mimicking the lower fidelity of a non-DNS simulation. Lastly they were sampled from the original $348^3$ down to $116^3$ in resolution.
 * __Model Input:__ low fidelity velocity (*u*), magnetic field (*B*), and their respective derivatives at a set time step in the future.
-* __Model Output:__ velocity stress tensor components ($\tau_{i,j}$) at the matching time step in the future, which effectively represents the difference between large and small scale structures of the system.
+* __Model Output:__ velocity stress tensor components ($\tau_{ij}$) at the matching time step in the future, which effectively represents the difference between large and small scale structures of the system.
 
 In this case, the probability density functions are overall consistent, with minor disagreement at the positive outliers, even though the prediction is performed far into the future (time = 9.48 ms, end of the simulation time). Predictive advantage is highlighted when compared with the analytical Gradient model that misses a large portion of positive data.
 
-![Predicting turbulent stress-tensor component ($\tau_{xy}$) in a core-collapse supernovae (CCSN). The model has been trained on a selection of dynamically evolving turbulence timesteps during the first 5 ms (out of the total $\sim$10 ms) of a 3D MHD direct numerical simulation (DNS) after the shockwave bounced off the core in a CCSN scenario. On the left, the two figures are the 2D slices of a 3D $\tau_{xy}$ prediction, with the right plot presenting a comparison of PDFs of the original 3D data, 3D ML prediction, and a conventional Gradient subgrid model.](ccsn_t160_tn1_slices_pdf.png)
+![Predicting turbulent stress-tensor component in a core-collapse supernovae (CCSN). The model has been trained on a selection of dynamically evolving turbulence timesteps during the first 5 ms (out of the total $\sim$10 ms) of a 3D MHD direct numerical simulation (DNS) after the shockwave bounced off the core in a CCSN scenario. On the left, the two figures are the 2D slices of a 3D $\tau_{xy}$ prediction, with the right plot presenting a comparison of PDFs of the original 3D data, 3D ML prediction, and a conventional Gradient subgrid model.](ccsn_t160_tn1_slices_pdf.png)
 
 # Acknowledgements
 The development of ``Sapsan`` was supported by the Laboratory Directed Research and Development program and the Center for Space and Earth Science at Los Alamos National Laboratory through the student fellow grant. We would like to thank DOE SciDAC for additional funding support.
