### 2021-11-23 80a6e60
Fixed compatibility with the latest opencv-python
	- 2D box filter is now working correctly
	- requires opencv-python >= 4.5.4
