+++ requirements.txt
@@ -8,7 +8,7 @@ matplotlib>=3.3.2
 mlflow>=1.20.1
 notebook>=6.4.3
 numpy>=v1.19.2
-opencv-python>=4.5.1
+opencv-python>=4.5.4
 pandas>=1.1.0
 Pillow>=8.1.0
 plotly>=5.2.0
