### 2021-11-23 5d6054e
Compatibility, 2D DS, quality of life (#138)
* Fixed compatibility with the latest opencv-python

	- 2D box filter is now working correctly
	- requires opencv-python >= 4.5.4

* Updated Dynamic Smagorinsky model

	- improved initialization
	- added 2D support

* Removed redundant parameter requirement in Eval

	- 'flat' condition is now passed along with
	  other input data parameters
