### 2021-11-23 #3199 @pikarpov-LANL
@dfm , thank you for checking in! Unfortunately, due to a family emergency on the other side of the globe, I had to spend a few weeks with my family. This caused a number of delays on all work fronts, but I just got back and trying to catch up.

That said, I think that I am happy with the paper at the point. All of the figures have been updated to reflect the latest state of Sapsan. In addition, the examples in the paper better showcase various capabilities of dealing with 2D and 3D data, numerous analytical tools, and compatibility with different ML libraries. Lastly, I made a few minor tweaks and have pushed the newest version of Sapsan. Here is the updated info:

Latest version: Sapsan v0.4.6
DOI: 10.5281/zenodo.5720254

Please let me know if there is anything else that I can do.
