### 2021-11-26 #3199 @Kevin-Mattheus-Moerman
I proofread the paper and reviewed the ZENODO archive. All seems in order. I will now proceed to accept this work in JOSS.
