### 2021-12-01 #3199 @pikarpov-LANL
@dfm , @MilesCranmer , @kburns , @Kevin-Mattheus-Moerman - thank you all for helping improve Sapsan framework and the paper. The review process itself has been immensely helpful for myself and my collaborators. This publication will make it much easier to promote Sapsan in the community. Lastly, it will be proudly highlighted in my upcoming dissertation!
